Limit instant app access to clipboard
Instant apps can access the clipboard (both read and write) in order for
copy/paste UI to work correctly. In order to limit the exposure of user
data Instant Apps can only access the clipboard while running in the
foreground.

Longer term it would be nice to gate Instant Apps clipboard access
around user interaction to prevent shenanigains, but this is a good
start.

Test: Manually verified that ClipboardManager.getPrimaryClip returns
null when the instant app transitions out of the foreground
Bug: 34231507

Change-Id: I7ea83e503d249c61e9df2d8bd83028e24705be11