EUICC API changes to support error resolution.
-Implement EuiccManager#startResolutionActivity. This is just a
pass-through to start the LUI dispatcher.

-Add and implement EuiccManager#continueOperation. The LUI calls this
method upon completion of a resolution action to continue the original
operation.

-Remove the "consentGranted" bit from DownloadableSubscription. This
was originally to be used by the LUI to communicate user consent, but
it is impossible to allow the LUI to modify the given
DownloadableSubscription while also preventing it from tampering by
the untrusted app. It's also unclear what kind of "consent"
was granted - it could be to disable the physical SIM, or to
grant an unprivileged app permissions to switch. Instead, this bit is
passed by the LUI to continueOperation() as an extra and propagated
down as an explicit parameter to the LPA.

Bug: 33075886
Test: Manual verification using test LPA
Change-Id: I17d8028e0ecea365ca5a2cc17424fc9645f5a489