Update unit tests for forward delete
Changes in the definition of grapheme boundaries for emoji (which
now relies more on what actually ligates in a font) need updating the
unit tests. The changes in the unit tests are cases where the old
code in Minikin assumed a single grapheme, but the new code looks in
the font and finds that there are two grapheme clusters indeed.

Change-Id: Icd3d2563776fb295f48e70b67a8af8e6bd3f4b38
Fixes: 37769578
Test: updated tests pass