Improved the media color extraction algorithm
The color extraction algorithm is tuned further:
We're now regarding more colors as white when
it comes to the background selection.

For muted color swatches we also factor in the popuplation.
Finally, we also factor in the mostdominant color to overtake the text
color in case it is very dominant.

Test: runtest systemui
Bug: 38168152
Change-Id: I858017b83559255ead48c3ab6c3bc16ed110e57c