Relayout before creating surface
Since we create starting window surfaces immediately (instead of
waiting for app visibility), we hit a weird corner case in which
the surface was created before the window was laid out. Thus, a
little bit later, we had to call SurfaceControl.setSize, which
made the whole transaction blocking, meaning that we blocked until
the next vsync before returning to the client (starting window).

Fix this by running a layout if the window wasn't laid out yet.

Test: Open hot app, see latency go down by an average of 8ms (up
to 16ms better!)

Bug: 32668632
Change-Id: I9817afc72127e32d7f8f4e497e0dccd7f180a9c0