Fix broken LoaderLifecycleTest#loaderIdentityTest
Bug 37994415

We have a preemptive check for getLoaderManager() to ensure that
the host is set before returning a value. In this test, the host
was not set because the fragment was restored and on the back stack.

The solution is to pop the back stack to ensure that it is added
so that it comes up to the proper state before getting the loader
manager.

Test: ran the broken test
Change-Id: Ic91cc5ca9468da22bdfef28cda1d4488b2f737f0