Change SlashDrawable semantics to fix QS crash
SlashDrawable wraps a reference to Drawable in order to store the
intended Drawable and optionally draw a slash over it. As such, it
should probably have similar semantics to Drawable, in that it either
exists or doesn't; if you set SlashImageView's drawable to `null`, you'd
expect that the drawable goes away, not that the SlashDrawable's
contained Drawable goes away but the slash stays.

Test: runtest systemui
Change-Id: I34b92716cc60ba5043ab68ae09954f5a5e97e562
Fixes: 62525439