BT device renaming: move edit icon and add metrics proto
On the bluetooth device details page, we need to have a dialog for editing
the display name of a paired device. This change moves the icon we need for
the menu item that launches the dialoge from SystemUI to SettingsLib (so we
can then use it in packages/apps/Settings), and also adds a metrics proto
entry to be used when the dialog is opened.

Bug: 62535241
Test: should be covered by existing tests
Change-Id: I63973e59cbd2eeac9bea04f66588f4e0b0254785