Fix QS wifi accessibility announcements
The state was being inverted in events to try to get clicks correct,
but events can occur at other times, so really we want to invert
the state between when the click happens and when the next state
change occurs. Also add a toString to fix equals checks on the content
description to make the announcements less noisy.

Test: manual
Change-Id: Ie4159664e9a81fd1719dbf7c06fba5e7b827cc6d
Fixes: 62679498