Respect the locale in vertical location of AM/PM
If the locale puts the AM/PM before the hour and minute in time
formatting, make sure TimePicker respects that by putting the AM/PM
above the hour and minute in landscape mode, where the AM/PM used to
always appear below the hour and minute.

Two sets of locales are affected. All of them put the AM/PM before
the hour and minute, but the first set also defaults to 12-hour, so
the change is most visible in these: Ewe, Chinese (both Simplified
and Traditional, including Cantonese), Korean, and Tamil. The second
set, Hungarian, Japanese, Myanmar, Sinhala, and Turkish, would only
see the difference if the time picker is forced into 12-hour mode.

Test: manual (tested en-US, ar-EG, ta-IN, zh-Hans-CN, ko-KR)
Change-Id: Ic5840d46d81a004294bda4644df02c721152686b
Fixes: 63352170