Internationalize ZoneGetter properly
Previously, the string for "GMT", as well as the digits and the
positive and negative signs for timezones were hardcoded (these
should have been different for various locales that use native digits
or an actual minus sign). Also, the number of digits used for the
hour was hardcoded to be two, where for Czech and Finnish they should
only use one digit if the hour is less than ten.

Fixes: 63675397
Bug: 63652836
Test: make -j RunSettingsLibRoboTests
Test: make -j RunSettingsRoboTests
Test: Checked time zone picker in Settings under en-US, fa-IR, ar-EG,
      cs-CZ, he-IL, sq-AL, and fr-FR.

Change-Id: I65191429b8b562f77977b1aa4a7db86a779dddc8