Fix ScrollView coretest
Due to another bug, adding new focusables dynamically does NOT
emit focusableViewAvailable. This caused the ScrollView to stay
focused despite having focusable children.
This caused one of the FOCUS_DOWN operations to be dedicated
towards sending focus from the scrollview itself. This change
sends focus to the first button before initiating all the DOWN
presses.

The change that caused this regression was allowing initial
focus to go to a zero-sized view (scrollview). This is NOT our
eventual desired behavior; however, since we allow zero-sized
views to be focusable, it didn't make sense to restrict it for
this one case.

Bug: 64096278
Test: testArrowScrollDownToBottomElementOnScreen
Change-Id: I8d6cc68a71bae630c02947aaf4dda15c27087422