Track update count of TSMS#mSpellCheckerMap
TextServicesManagerService#buildSpellCheckerMapLocked() is known be
one of the biggest contributors to CPU and I/O bandwidth consumed
by TextServicesManagerService (TSMS). To check when it gets called
and to see how a CL would affect its call count, it would be useful
if we can easily get the call count of that method.

With this CL, how many times TSMS#mSpellCheckerMap was rebuilt will be
included in the dumpsys.

Bug: 64449660
Test: adb shell dumpsys textservices | grep "mSpellCheckerMapUpdateCount="
      to make sure the count is increasing with each switch between
      users.
Change-Id: Ia1f50c756c332b3329b6e28e8edb7673cf9124f2