Add new "fast" filter for ambient light.
The slow filter wasn't converging fast enough when we shortened the
debounce timings but we still want it to ensure there's a long term
change happening, so add a new filter that converges much faster. This
new fast filter is just a weighted sliding window just like the slow
filter, but it has a significantly smaller horizon.

Bug: 64514692
Test: manual
Change-Id: I1e990ac1e6a56268fa42a85399a2d41cc8dd2104