Ignore only internal volume keys when in VR mode.
The concern here is that the elastic strap on some VR headsets will
press the volume keys on the side of the device. Given that, external
keys (e.g. bluetooth headsets, USB keyboards) aren't a concern and
should be allowed to go through.

I ran a couple quick traces and the initial fetch of the InputDevice
takes ~0.1ms and each cache lookup following that takes ~0.01ms. This is
probably okay to happen on the dispatcher thread since it's both rare
and reasonably cheap, but if we find this to be a performance bottleneck
in the future we could just add a flag to the input event itself
denoting whether it comes from an internal or external input device.

Fixes: 63909929
Test: flashed change, put the device in VR mode and saw that the side
      button volume keys didn't work but a USB-C headset's volume keys
      did.

Change-Id: If3dfcd374abd79236c2c38e5b9439564c6b49ec8