Allow translucent RenderNode->Bitmap captures
Set translucent to true when doing createHardwareBitmapFromRenderNode.
This ensures that the buffer is initialized to transparent which
is a safer default for this path. Marginally less performant in some
cases, but as this is a rare capture safety is preferred for and
performance can be revisted if it turns out to be an issue.

Bug: 65164016
Test: builds
Change-Id: I6467c0ccf46e5e2a87c31cc4855ced8275236ca9