IP connectivity metrics: fix LINGER/UNLINGER logging
This patch fixes a regression introduced by commit 97bfd276fb008
for logging NetworkEvents when lingering and unlingering a network.

Commit 97bfd276fb008 removed an overloaded constructor for the
NetworkEvent class, which caused NetworkEvents with event type of
LINGER or UNLINGER logged in ConnectivityService to have incorrect
event types (set to the network id instead) and incorrect duration
(set as the event type instead).

Bug: 34901696
Test: runtest frameworks-net
Change-Id: Iab97a58ca805413617c8e8b4553404625a820ceb