Add hooks to configure system button text styling.
Adds configuration values to change the styling of buttons on system
surfaces. Specifically, adds config_fontFamilyButton and
config_buttonTextAllCaps to control the default font and whether text
is all caps.

Additionally, adds plumbing so that these configs are utilized by
Theme.DeviceDefault (and descendent) themes. Some styles are added/updated
for this support.

Bug: 64122537
Bug: 69314526
Test: ag/3228846
Change-Id: I08406633bd744c6f58d5144d035aa27bf9c175df