Always show window divider to accessibility
Accessibility typically can only see windows that the user
can interact with. This CL carves out the split-screen
divider as an exception. The reason is that services can
use the presence of the split-screen divider to know if the
device is in split-screen mode. But the divider becomes
non-touchable if an IME appears in split-screen mode.

Services like TalkBack need some way to tell for sure that
they are in split-screen mode. Absent a new API for them
to query, simply making sure they can see the divider seems
like the simplest way to provide this.

Bug: 28768841
Test: A11y service CTS
Change-Id: I92b89c842689e4c4c69a732490150ca032de46cc