Handle z-layering in animation layer
We use the prefix order to determine base layer within the
animation layer. This ensure that for the animation windows, the
z-ordering during animations doesn't change.

We then boost anything that needs a z-boost to 800570000 + prefix
order, such order within the boosted layers is preserved as well.

Also fix an issue where the thumbnail wasn't attached to the
animation layer.

Test: WindowContainerTests
Test: go/wm-smoke
Bug: 64674361
Change-Id: If5909bd87a12f1d8920c7232acab0f3d17be0f6c