Rename onException to onPartialImage
Bug: 63909536
Test: CTS: I8d1672180d8325ae1caf44f0bbf41036b94e6253

In ImageDecoder. The goal is to tell the client that there is a partial
image (which they can use if the callback returns true) and also allow
them to know why the image is partial.

Change how the return value is handled. Instead of returning null, which
is inconsistent with the rest of ImageDecoder, throw the Exception.

Change-Id: I56c38a624c978aa6e6d00fc927b5e355bf9c718a