Prevent getting from transit when interesting window is exiting
AppWindowToken determines when alldrawn state is done to then choose the
transit animation depending on if the next window has a wallpaper or
not. In terms of launcher, it has -1 window without a wallpaper while
the main window has a wallpaper. Meaning that sometimes quickly going
home, services thinks the main window of launcher is exiting and so it
is not considered interesting but the -1 window is drawn and interesting.
Then the system thinks going to launcher has no wallpaper and chooses the
wrong animation.

To fix this issue, assume that if there is a base window (not splash or
starting window) then there is always an interesting window. All the
other windows can enumerate the interesting count until all the windows
including the main window is drawn and ready. This will allow launcher
to wait for the main window to exit and complete drawing before choosing
the next transit animation.

Test: manual - see bugs below
Change-Id: I4839a6f706fa43d295524a5e252a62e8babd1e7d
Fixes: 65304663
Fixes: 65559439