Ensure that PostProcess Canvas is released
Bug: 63909536
Bug: 63908092
Test: CtsGraphicsTestCases (ImageDecoderTest)

Refactor a method for calling postProcess from ImageDecoder. This will
be shared with the animated drawable (TODO). Call
PostProcess.postProcess in Java inside a try block to ensure that the
Canvas is released. Otherwise, a client could hold on to a pointer to
Canvas and keep using it, even though we have removed its backing.

In addition, share code for calling nDecodeBitmap.

Change-Id: I81ce2befce91ac1e27c78ad059c4b173a2c7e679