Fix surface properties not getting applied correctly
When we switched to calling prepareSurfaces in
applySurfacesChangesTransaction, we switched from
computeShownFrameLocked to prepareSurfaceLocked, which has an
additional check of waiting for opening, which caused that a
surface size change from a layout never was applied.

The original comment on that check doesn't make sense anymore,
given that we don't animate in prepareSurfaces anymore, so we can
just go ahead and remove that check.

Test: go/wm-smoke
Test: Open voice input activity
Change-Id: I46664fd3c8f9a5c549fdb245125f05f290162ebf
Fixes: 71787613