Fix flicker with remote animations
There was a race condition where we notified the controlling app
already about transition start, which itself started the animation
and applied the transform of the first frame. However, a little
bit later, we applied the pending transaction from window manager,
which overrode the properties again, leading to a flicker.

Original CL: ag/3535475

Test: go/wm-smoke
Test: Press home button, observe never a flicker
Bug: Surprisingly there isn't one yet.

Change-Id: I84b2e0fd4dcd7a01687e18f428a8f900ed43d75f