WM: Split portrait and landscape status bar size
For devices with a display cutout, it is no longer feasible to have the
same status bar size in portrait and landscape.

Special care has to be taken because a lot of third-party apps (and bundled apps)
rely on @dimen/status_bar_height, so we make sure that in -land it is set
to the landscape value.

Note that resource overlays for @dimen/status_bar_height MUST be renamed to
status_bar_height_portrait, otherwise they will only override the portrait
orientation, while landscape will continue to be 24dp!

Bug: 73717865
Test: Enable display cutout, ensure status bar is smaller in landscape.
Change-Id: If2718e53c3a10e3cab092d4b6a0a2a3bdd0cea81