[Notif] Add Blocking helper to swipe
Implemented blocking helper UI as part of guts - the
ExpandableNotificationRow that's swiped away is effectively detached
from the notification logic but kept in as a view in the layout (it'll
still show up when looping through the normal view hierarchy). We
purposefully open up NotificationInfo with the intent of showing it as
the blocking helper.

It's animated away/closed when the user interacts with any other content
or swipes away, as with guts (Can be changed in the future if we want it
to stick around for longer).

KIs, that will be tracked in bugs once CL is in:
- Swiping a notification if it's partially hidden by the shelf does not
bring it up into focus
- If only one notification is showing, the shade auto-closes as if it
didn't have any notifications (which it technically doesn't)
- Animation on multiple blocking helper shows cuts too early and doesn't
allow the notification to transition out
- Long press on notification dismisses (it should probably ignore it)
- Metrics are still missing
- Accessibility needs a runthrough

Bug: 63095540,73783854
Test: Visually & Ran all tests
Change-Id: Iaa5e3d5537bbb72c946fdd7ee35b155e22ca2b05