Don't add stopping apps in relaunch based on stopped flag.
The activity was getting added to the list of activities that should be
moved to stopped. This was causing the window to get removed and then
re-added, which is where the flicker was coming from.

The check to add stopping apps when the flag r.stopped is set was a
fix for a lifecycle issue with relaunch. However, this issue is now
fixed without this. The stopped flag is not set appropriately because
activities that are paused can still be considered stopped if they never
went through completeResumeLocked, like in multi-window.

Change-Id: I2e49bfa031caa189cc1e440749d0a22b30b76792
Fixes: 75276931
Test: Tested original issue to make sure it still works from b/30060825
Test: Split screen no longer flickers when moving divider