Correct lookback for while-idle alarm rate limiting
Lookback for while-idle alarm "last delivered?" bookkeeping was clamping
at zero, but that means "when we rebooted" in the elapsed time base, not
a point in the deep past.  As a result, testing sequences that forced
idle mode would behave unpredictably on devices that were rebooted
recently.

Change the lookback defaulting such that the first "while idle" alarm
that triggers after reboot will always be considered deliverable.

Change-Id: I0a35e046ecf858cd44523a6cece4a6a0779194bc
Fixes: 77981603
Test: atest CtsAlarmManagerTestCases
Test: POC app