Generally improve brightness dumpsys.
For the recent spate of autobrightness bugs, being able
to determine both current state of the display brightness and the
overall situational context of the device (e.g. ALS readings) is
critical to quick diagnosis. This tidies up the logs and adds some more
state information so we have more context about the situation the device
is in.

Bug: 78909876
Test: adb shell dumpsys display
Change-Id: I0e3cd5d00268b5b44f3d486fae8940d7d7d2ed2e