Hide inputMethodClientHasFocus() from apps
inputMethodClientHasFocus() implemented by WindowManagerService is an
internal method that is used only from InputMethodManagerService.

For obvious security reasons, we do not need to expose this method to
application processes.  We can easily achieve this by simply moving
this method from IWindowManager to WindowManagerInternal.

Note that this is a mechanical refactoring.  There should be no
observable behavior change.

Fix: 112722651
Test: atest CtsInputMethodTestCases CtsInputMethodServiceHostTestCases
Change-Id: I20c618174d8279a9a57f458fb908235f452b2281