WebView: fix SafeBrowsingResponse javadoc bug
No change to logic, only minor javadoc change.

This fixes javadoc in SafeBrowsingResponse.java. Previously, the first
(and only) sentence did not end in a period (it ended with a quote
character). The Android javadoc parser doesn't handle this well--it
requires a period followed immediately by some whitespace character.

This wasn't a big issue because this was the only sentence. However, if
we ever add a second sentence, both sentences will appear as the
one-line "synopsis doc" for the methods.

See http://go/android-api-guidelines#heading=h.teqny5h9ohqc for details.

Bug: 115441402
Test: make docs
Change-Id: Ic363ed1ecf3d169d598ae85d55297d0a520a6604