Camera: Add support for recommended stream configurations
A camera device must be able to suggest power and performance
efficient stream configurations depending on the client
use case.
The new configurations should only be a strict
subset of the available exhaustive list.
Clients must be able to query the additional information using
familiar API which will not require extensive changes on their
side.

Bug: 64029608
Test: Camera CTS
Change-Id: Ib00394669d40e389426f5ba2f4c06061feaea340