Only inflate public view when notif is redacted.
This CL changes notification inflation behavior to only inflate the
public content view when the notification needs to be redacted in the
first place (i.e. the setting to hide sensitive notifications is on and
the notification would be hidden under the setting). This saves around
.1 MB per notification.

Bug: 111809944
Fixes: 111809944
Test: runtest systemui, manual tests (toggled hide sensitive on and off,
tested ambient pulses under both)

Change-Id: Ia1d00db00c5164f043c999578f0be4b55f223f8e