Not keep relative position if display changes.
Launch bounds already consider display rotation so there is no need to
apply transition to keep relative positions. We may need to add bounds
adjustment when we detect display changes to make sure the task is not
completely out of new display bounds, but it's not in the scope of this
CL.

Bug: 118184707
Test: go/wm-smoke. Launch freeform tasks on a display with rotation
works as expected.

Change-Id: Id0c3a3546b279de4170cb62f39f9825904267449