Add owners to the Shell app
The Shell app has elevated privileges as it is used for development
and testing behind the USB debugging developer option. This change
adds owners file to control the changes to the Shell app and more
specifically to its manifest file to ensure only sensible singature
protection permissions are granted to the shell.

Test: manual

bug:80415658

Change-Id: Ic7a442320f052d3c698ed923f1a953e4bee646eb