Add support for illustrations to SimAppDialog.
Previously these were provided in build-time resource overlays by
overriding the entire install_carrier_app_activity.xml layout. This is
no longer feasible with the move to runtime overlays as the layout
cannot be precisely mimicked due to depending on styles and other
parameters from the base APK (and is generally a bit more brittle).
Instead, pull in the needed container from the current overlay, add a
stub image and content description and hide the container by default
so the base APK can be compiled, and provide a boolean that can be
overridden along with the image and content description that is used
to toggle the image display at runtime.

Bug: 118823052
Test: atest GtsSimAppDialogTestCases (with and without overlay)
Change-Id: Ic706df3091ce6c9f5d64ee22da6fa61b9b7b1e3c