Remove com.android.i18n.phonenumbers.... classes from light grey list
Removes almost all the com.android.i18n.phonenumbers.... classes from
the light grey lists as they are now using UnsupportedAppUsage
annotation.

The only ones that are left are ...->values()... methods which are the
synthetic methods created by the compiler when compiling an enum class.
It is not currently possible to use annotations for those as they do
not exist in the source code, bug 119861512 tracks the changes.

This was tested by making and then manually checking that the generated
out/target/common/obj/PACKAGING/hiddenapi-light-greylist.txt was the
same (after sorting) before and after this change.

Test: see above
Bug: 117818301
Change-Id: Ifb8f243ca2e519fd9d0aca04722ae5155913e91e