Fix ActivityView input forwarding.
The current behavior for input expects the input to be in screen space.
However, ActivityView was forwarding input in View space so it was
causing the touch to be in the wrong place.

This fix updates the touch event so it's translated into screen space
before forwarding to the InputManager.

Test: ActivityView touches
Change-Id: I8b86a17adb166fdd95f24610f2ba9ae55c39f7c5
Fixes: 111440225