Remove foreground user check in shell commands for IMMS
In general adb shell command takes an optional user ID parameter when
we want to request something for a certain user.  With that, checking
whether the caller is a foreground user or not does not make much
sense when handling shell commands.  The entire point is whether the
caller os such shell commmands has sufficient permission or not.

Bug: 34886274
Test: 'adb shell ime reset' still works
Change-Id: I0aca79a33440f3636990ed9ef0efbbe0e4a42f77