Grant removed install permissions
Historically removed permissions just stayed forever in their default
state. E.g. install permissions are granted and runtime permissions are
granted for pre-M apps. In Q we should keep this behavior. Still we have
to make sure that we never set the review-required flag as this flags
will never be reset prevents the app from starting.

Fixes: 120981154
Test: - Manual verification of review-required flag for pre-M app
        - when freshly installed
	- after permission review
	- after permission reset
      - atest android.appsecurity.cts.PermissionsHostTest#
              testCompatDefault22,testReviewPermissionWhenServiceIsBound
	while set-isolated-storage was set to true
      - atest android.content.cts.ContextTest#testCheckPermissionGranted

Change-Id: Ic27c17d2e93f6e2f87127cd4aec7281e6fbdbb8f