Add API to check if activity can be started on a display
Some activity launch restrictions apply to virtual displays, which
are not communicated to apps. Currently there is no way to check
this in advance before starting an activity. This means that an app
get an unexpected SecurityException after calling startActivity and
therefore cannot know when to show in their UI a possible option to
launch on a secondary display.

This CL gives adds a public API to check the possibility of launch
on a specific display.

Test: ActivityManagerMultiDisplayTests
Bug: 119575501
Change-Id: Ieb70f0bb79b1a88b7284a19af2efeeb1fcb90f75