Use rotation to report handling for orientation.
Before this CL DisplayContent relies on whether DisplayRotation respects
app requested orientation to report handling, but it may not work if the
display respects app requested orientation but it actually doesn't need
to rotate itself for the new requested orientation. For example, A
background task could be letterbox-ed if the foreground task doesn't
lock the display rotation, and if the background task changed its
requested orientation we might not change its letterbox-ed state.

Fixes: 122640019
Test: Manual tests.
atest WmTests:DisplayContentTests
atest WmTests:TaskRecordTests
Change-Id: I36e6bbdc48a479e860cf4c834367ed1efbb1e9ed