Add config_defaultRoleHolders for specifying default role holders.
This config string array will be an array of roles and package names
of their default holders, with each item in the format of "ROLE_NAME:
PACKAGE_NAME_1, PACKAGE_NAME_2". This config will be read by
PermissionController to initialize the roles. This change also fixed
an issue preventing role controller to set default holders correctly.

Bug: 110557011
Bug: 122730135
Test: build
Change-Id: I90ffd156d26098235c821e7937950c0b420e0e7f