Prevent reallocation of Surface object in copyFrom
Previously when we sent the Surface to the client (as opposed to
the SurfaceControl) Surface::readFromParcel would take care of reusing
the object if the underlying IGBP had not changed. Now that we send
the SC and copy it in to the Surface, we need to add similar logic
in Surface::copyFrom.

Bug: 122588130, 122659537
Test: Manual
Change-Id: Id0c8e61a9aff9b3c37b2a783717c15df3b36f175