Fix deferring making activities active
New logic for checking whether activities should be made active did
not take into account deferring resume during launch. This lead to
making activity resumed before actually launching it, and
top-resumed state change callback was ignored on the client side
since activity was not registered on the client yet. This in turn
made the client think that it was never made top-resumed before
loosing the top-resumed position.

Bug: 123587738
Test: atest WmTests:ActivityRecordTests#testShouldMakeActive_deferredResume
Change-Id: Ieb3b98b4ddb4819805fa0e818ecf8359487eaaff