Trigger CtsDevicePolicyManagerTestCases in postsubmit
In this configuration the tests will only be triggered by changes
to services/devicepolicy/** and core/java/android/app/admin/**, so
we won't be able to catch breakages coming from the rest of the tree
with a single CL precision. But let's see which parts are causing
the most breakages and add them later.

Bug: 124653467
Test: atest --test-mapping core/java/android/app/admin:postsubmit
Test: atest --test-mapping services/devicepolicy:postsubmit
Change-Id: I58f87fff4b5b5e3e4fb8a333dd6ac4dd4100b587