RollbackManager: Backup and restore split apks.
Previously RollbackManager would only backup and restore the base apk
for a package. This change fixes it to backup and restore split apks as
well.

Adds a new test for rollback of packages built with splits.

Also, switch to using resources to distinguish between v1 and v2 of an
application instead of meta-data, given we use resources anyway now to
distinguish between split versions.

Bug: 124848643
Test: atest RollbackTest, with new test for splits added.
Change-Id: Id1f32db7b728bdeda6c7223bc69a7e80ba569dda