Count letterbox into non-magnified bounds
On device with display cutout, apps could decide if their windows
are allowed to extend into the cutout area. If not, a letterbox will
be applied such that the window does not extend into the cutout. The
letterbox area won't be magnified so we wouldn't draw magnification
overlay there.

In this case, the magnification overlay would draw around the entire
screen when there is a heads-up notification or another bigger
magnifiable window becomes visible on top of letterbox area at that
moment.

However, we don't really count the letterbox into non-magnifiable
area which would be counted as dirty rect. Hence, we only redraw
smaller dirty rect. The old bigger boundary is outside the dirty
rect and still in the surface.

Bug: 79117529
Test: a11y CTS & unit test
Test: manual
Change-Id: Ie1cee78713016716fa1a203ffd12cf1612329590