Fix javadoc HTML escaping.
HTLM special characters ('<', '>' and '&') don't need to be escaped
inside {@code}, even though they do elsewhere in javadoc. And in fact
they shouldn't be - because if they are escaped then the escaped
version shows up verbatim in the developer docs.

See for example:
https://developer.android.com/reference/android/graphics/drawable/GradientDrawable
https://developer.android.com/reference/android/support/v13/view/inputmethod/InputConnectionCompat

This change just removes the escaping to fix the public documentation.

Test: N/A (only comment lines changed)
Change-Id: I131b3b3e26547e30cff7b67514d177c48926bb26