Protect against null Displays.
Because the display callbacks happen asynchronously, if a display is
added or changed and then immediately removed it may no longer exist by
the time we query DisplayManager about it. This results in an unexpected
null Display object, causing DisplayModeDirector to crash. We guard
against this case, ignoring the display until it gets re-attached.

Fixes: 127875518
Test: manual
Change-Id: I32acd77a095181155dbcb57922f616b2ac57192c