Aligning job readiness checks.
areComponentsInPlaceLocked fell out of sync with
isReadyToBeExecutedLocked. The former is used to determine if constraint
exceptions should be done for a job. The latter is used to determine if
the job can be added to the pending jobs list. If they're out of sync,
we may do a little bit of unnecessary work.

Bug: 125505588
Test: atest CtsJobSchedulerTestCases
Change-Id: Ie11daa1dad7a97ff6a582706db8b4ca38f53a76e