Remove keyRepeat from KeyButtonView
This attribute is not used by anyone in Android master. It therefore
always defaults to "true".

There are some potential usages in the "experimental" project, but it
will be up to the owners of those copies to clean this up.

Bug: 128553584
Test: presubmit
Test: search for "keyRepeat" in Android via code search and examine
every usage (166 results). Also grep the local tree for "keyRepeat"
inside the frameworks/base and the vendor/unbundled_google projects.

Change-Id: If2bc354ff2a2259874dc19a973e1e0152d9fd9c3