Add count to TouchEventReported atom
We might have to reconstruct mean and st. dev. over arbitrary periods of
time and across multiple devices. Currently, we are just looking at the
average and st. dev. values. In order to be able to reconstruct, we need
to know the number of events in each chunk that is received.

Add the needed field here.

Bug: 111431676
Test: out/host/linux-x86/bin/statsd_testdrive 34
Change-Id: I5aa4b759404308b98d4e6b5ed5a94e4c760dc447