Implement ExplicitHealthCheckService in ExtServices
If658567ca41bb0328e279735897fc50ab76abc11 added the
ExplicitHealthCheckService API to allow ExtServices to request explicit
health checks for packages and report to the PackageWatchdog.

We now implement this service and add a rudimentary NetworkChecker
that relies on listening to the ConnectivityManager for validated
network requests.

Bug: 120598832
Test: Manually tested that after an APEX update, the network stack
does not pass the explicit health check until WiFi is connected
successfully. If Wi-Fi is never connected and the network stack
monitoring duration is exceeded, the update is rolled back.

Change-Id: Ifa968b4c7107cabedf186dffbc5eb771a3840e7f