Remove the pending ARC termination Action when boot up and init ARC regardless of the current ARC status.
Note that when the current device goes to standby, it terminates System
Audio Mode to hand the audio to the TV. But some TV might send out ARC
Termination message after we going to standby and we are caching these
messages to handle later.

This causes Terminate ARC being handled when the device boots up and
sends out incorrect ARC Termination messages when it's supposed to init
the ARC.

We move the Arc Terminating Action at address allocated here and init
ARC regardless of the current ARC status in case the TV and the current
device's ARC status are not synced up.

Test: manual
Bug: 130199965
Change-Id: I15854802efa999814f27633dfcc83c0101a5ad59
(cherry picked from commit 119fb0e8b41d2b2e85e16b9126cf88002577f180)