AccessibilityEvents for visibility changes should be subtree events.
We don't report hidden views to the AccessibilityService (see ViewGroup.java).
Therefore changes to visibility are changes to the Accessibility Tree, and should be treated as such.
Because AccessibilityEvents aren't sent for hidden views, previously, hiding of views wouldn't be reported to the accessibility service, which is a pretty bad bug.

Test: Added CTS Test in another CL. Played around with this build for a while.
Bug: 130273130
Change-Id: Ie2021f1926bfe9e67e9aaa25a12c2e7cbfc74500