Revert "MemoryIntArray: don't double close."
The reverted change causes a regression where we can get an IllegalStateException
during finalization as we are adopting the native fd in a ParcelFileDescriptor
which takes ownership of the fd. However, the order of finalization is undefined
and if the ParcelFileDescriptor is finalized before the MemoryIntArray we would
get an exception when running the finalization of the latter.

This reverts commit 89dfbfa269d32534a0b110593a2ed7e7183d21a7.

bug:124056170