Revert "MemoryIntArray: track the owned file descriptor in a PFD."
The reverted change causes a regression where we can get an IllegalStateException
during finalization as we are adopting the native fd in a ParcelFileDescriptor
which takes ownership of the fd. However, the order of finalization is undefined
and if the ParcelFileDescriptor is finalized before the MemoryIntArray we would
get an exception when running the finalization of the latter.

bug:124056170

This reverts commit c81f53f7f114962f757312ac884a279035fe0584.

Change-Id: I8debb9c5f4c87b1a657084139b27f40b7956fe59