Fix error source of system server crashes
System server crashes have a null ProcessRecord. This meant that we
would always annotate the error source as UNKNOWN for system server
crashes. Now, the check for error source first checks if the process
name is system server, and if so, makes the error source SYSTEM_SERVER.
Otherwise, it continues to check the ProcessRecord.

Test: manually mocked a system server crash and ensured it was logged
properly
Bug: 130438842

Change-Id: I71fc9c3ecdd3298ab878b1b14faae2567b5e640a