Fix always-null children on NotificationEntry
Looks like _someone_ decided to make an array of children and _never_
poplate it.

What used to happen was that callers into NotificationEntry would grab
the row and call ExpandableNotificationRow#getNotificationChildren() on
it to iterate the children and call things (notably setKeepInParent()).
During the work to refactor NotficationEntry into its own class,
NotificationEntry#getChildren() was created so that we could iterate the
entry's children instead of the view's children. Unfortunately it was
never populated.

For now, just call through to
ExpandableNotificationRow#getNotificationChildren() and map those to
their entries.

Fixes: 130018576
Test: atest SystemUITests; visual
Change-Id: Ica629b96c8fa2ee2fb45fdf588ee7d1efc836883