Add timeout for displaying 5G icon
This change add an optional filed that allows carrier to change the
timeout.

System UI will show the 5G icon and start a timer with the timeout
from this config when the device connect to a 5G cell. System UI stop
displaying 5G icon when both the device disconnect from 5G cell and
the timer is expired.

Bug: 136107473
Test: build
Change-Id: I9720ec127865daa03fb2c23fbdd7d4ebf5e30e1b