Switch from android.text.format.Time
Convert a use of android.text.format.Time.format() to use calculations
based on java.time instead. This avoids future Y2038 issues associated
with Time.

Note: The new code doesn't support digit localization for, for example,
Arabic devices. I strongly suspect that the digit localization would not
be intentional in this usecase so it has not been preserved.

Bug: 16550209
Test: build / boot / treehugger
Change-Id: Ia246e63a1cd9be08f78f54c38b45c64e4795618a