Use REVOKED_COMPAT instead of MODE_DEFAULT when syncing app ops.
To get rid of yet another permission state stored in app ops.

MODE_DEFAULT was used to record that the app op is set to a denied
mode by restricted permissions, in contrast to the MODE_IGNORED set
when user denies a permission. However since we have the
REVOKED_COMPAT flag now, we can just use MODE_IGNORED and get rid of
the permission state stored in app op and related logic for it.

Bug: 136503238
Test: manual
Change-Id: I2199188d7feeb37359ad2cb5e7621a3930488e5f