Try to rebind disconnected wallpaper service for 10 seconds.
When we received the wallpaper service disconnected, we could not
immediately know the reason for the disconnection, so if the service
was disconnected because the package was updating, and we received
the update broadcast after 1 second, in original design we will only
revert to the built-in wallpaper. However, it will be too late to bind
the original service after we receiving the package update broadcast.

In this CL, if currently wallpaper service is accidentally disconnected,
we will try to re-bind it for 10 seconds.

Fix: 138973259
Test: Manual install wallpaper service and delay the package update
broadcast on purpose, verify the service can be bind back.
Test: atest WallpaperManagerTest

Change-Id: I11891098164372002b36c558f670621d8c852af8