Removed FLAG_EPHEMERAL from pre-created users.
When a User is pre-created with that flag, its CE & DE keys are not persisted on storage. Hence,
when that pre-creatd user is converted to full-user, the expected keys don't exist and the user
doesn't boot properly.

This change fix the issue by making the pre-created user non-ephemeral, and changing it to
ephemeral when it's converted to a "real" user.

Test: manual verification on Automotive by creating a new user than deleting it using the UI -
that creates and switches to a new guest user (which was pre-created); without this change,
that user  would be hanging with a "Android is starting..." message

Bug: 143092698
Bug: 143553590
Change-Id: I2699f55a30d72fd5a88202bf6c83f7ca7c65fac0