Fix double-add of calls
In the api compatibility logic for the new call states, we might end up
adding a call by accident in internalUpdateCall if a call was updated in
the NEW state, and it'd then get added again when internalAddCall gets
called.

To fix this, restrict the call states that would result in an add from
internalUpdateCall, and also don't add the call again in internalAddCall
if it's already present.

Fixes: 143049799
Test: CTS
Change-Id: I2a6b646a5200fedf9e02029a7120595210a24c76