Start FUSE sessions before Vold mount returns
Before I45238a31df71286f67ef1c65c711d0085d72e97f, after a vold mount,
we could not get the mount paths immediately. So we had to wait for
mount state change to MOUNTED before starting the FUSE sessions.

Now, we pass a callback to Vold#mount to retrieve the fd and paths for
the FUSE session. This blocks the mount call until the FUSE session is
started.

Test: atest AdoptableHostTest
Bug: 144275217
Change-Id: If441c9782596f0e7b963ad10571ce2fa19746ee6