Use framework-statsd-stubs in frameworks/base/Android.bp
The non-updatable part of the platform does not use framework-stats, but
framework-stats-stub. In doing so, some APIs in StatsEvent were made as
SystemApis.

FrameworkCoreTests was also modified to avoid the implicit linking with
'framework.jar' that doesn't have the test-only APIs for statsd. To
disable the implicit linking, sdk_version was set to core_platform and
the dependency to framework.jar was replaced with framework-all.

Bug: 146167933
Test: m
Test: m FrameworkCoreTests
Change-Id: Ic447993a13f74c09fabb04fe32b9b94d78a34ce2