Change provideContextImage() API to accept Bitmap
provideContextImage currently takes in taskId which works for the
Recents screen.
In Android QPR2 timeframe, content suggestions also powers Screenshots
feature which requires passing an explicit Bitmap.
This change adds a Bitmap parameter to the 'provideContextImage'.
Launcher/Recents will continue using the taskId-based API to obtain a
fresh snapshot using the Activity/Window Manager; or obtain an on-disk one for paused tasks.

Bug: 147324195
Test: atest ContentSuggestionsManagerTest
  in separate CL ag/10146598
Change-Id: I686eb8d90a4a44fdd80403ef2df58020c82ca69a