Add ApexContext API including methods to get paths for data directories.
This adds a new ApexContext class as a system API. Methods are included to
obtain paths to the APEX data directories.

In future the aim is for this ApexContext to be provided to code running
within an APEX automatically, so that data directory paths are available
only to the APEX that should access them. For R it is not feasible to
implement this, so for now a factory method is included to create an
ApexContext from the apex (module) name.

See go/apex-context-r for more information.

Bug: 141148175
Bug: 142484007
Test: ApexContextTest added
Change-Id: Iaf90e12383aa219a73a7818dab6cde6a850597ba