Allow the visual dismissal of foreground service notifs
This CL adds support for a new feature flag
(notifications_allow_fgs_dismissal) that enables a user to visually
dismiss a notificationg that represents FLAG_ONGOING_EVENT or
FLAG_FOREGROUND_SERVICE into a new bottom section of the shade. Tapping
on the list items in the new section reinstantiates them back to the
shade.

In this CL:
- the new device_config flag + feature controller for it
- ForegroundServiceSectionController that can track dismissed FGS
notifications and then show them in its own decor view
- ForegroundServiceDungeonView that is optionally added at the bottom of
the shade
- Changes to the logic for NotificationEntry#isClearable to take into
account the new flag

Todo:
- Minify the dungeon view before first tap
- pin the dungeon view to the bottom of the notification shade

Test: atest SystemUITests
Test: adb shell device_config put systemui notifications_use_people_filtering false && adb shell kill $(pid com.android.systemui)
Bug: 136562823
Change-Id: I8359e8c3ad9de89cd867f5ac147d62e22b9cf816