Use correct actor string for OverlayableInfo
Need to use second, which is the actual value. Was using
the key and thus the overlayable name and tests were
failing trying to validate against this invalid actor
string.

Also moved NO_NAMED_ACTORS failure down a step, as that
was masking the core issue.

Also adds the missing overlay tests to the
com.android.server.om TEST_MAPPING. The imports tag
was not used to avoid importing all of the core tests.

Bug: 145923978

Test: atest android.appsecurity.cts.OverlayHostTest

Change-Id: I9cb790cf8966e7eaf7554feccf960a3debb1a0c3