Remove NetworkClass from TelephonyManager
NetworkClass isn't especially useful because the APIs for
getting and setting preferred networks have been converted
to bitmasks of individual RATs.

Removing this because it's causing confusion about the proper
structure of the API surface. There are < 100k total usages with
the vast majority (60% or more) coming from two apps.

Because NetworkClass is used in RadioAccessFamily for a sort of
comparison function, rewrite that function, porting and updating
logic from TelephonyConnectionService to that compare() method.

Bug: 128572818
Test: atest RadioAccessFamilyTest
Change-Id: I1f279fadd8c543b845fce0f13dc62c1480757692