Fix LogEventQueue_tests
These tests were broken by ag/10444161, which deleted a test-only
LogEvent constructor (the tests still built though because there was
another constructor that accepted the same parameter types but with much
different semantics).

Also fixes a bug in TestSlicedCondition that was introduced by one of
the later patchsets in ag/10444161.

Test: bit statsd_test:*
Bug: 149590301
Change-Id: If7de80922680c63505a4b30a1dcefdc47be13219