Remove haptic primitives that aren't ready yet.
We have some concerns about a couple of the primitives we're exposing,
so rather than risk committing to an API surface that we may regret and
aren't 100% confident in the implementation of, lets hide it and keep
them for private usage for now.

Also, rename LIGHT_TICK to TICK; we shouldn't expose the internal
differences to users since it just makes the whole surface more
complicated to explain.

Fixes: 151085071
Fixes: 151261619
Test: N/A
Change-Id: I2872e8f52202d6192de6aa7e3081778486174f4e