Remove StateProvider from LockIcon
Whenever the state changes in the system, we now push that information
directly to the LockIcon, instead of pushing what we think the prior
state was and letting it query for the current state. This allows
the LockIcon to properly track what _it_ thinks the old state was
(when it last changed its icon) and then react accordingly.

This has the added benefit of being easier to understand with
less code.

Fixes: 153117689
Test: atest SystemUITests && Manual
Change-Id: I8f2e3e1d2287b26ca44b7eb6df8de392efd4bb2a