Remove isGroup bind param from row content binder
What the content binder implementation actually cares about is whether
it should inflate a low priority content view or not, so it doesn't
actually really need to know anything about the group state. The
actual decision making logic for that should happen outside that
class, namely in NotificationRowBinderImpl.

Due to grouping being determined after inflation, however, we also have
to recheck at a later point in the pipeline to see if we inflated the
wrong view and correct it. We put both these methods into a helper class
for now until the new pipeline is implemented.

Bug: 153602376
Fix: 153602376
Test: atest SystemUITests
Test: Add min priority group in Notify APK and see that views are more
than just their header view when unexpanded

Change-Id: Iab7a1cd31a7d4114dec829e24c65cb04dde78667