Fix possible deadlock in incoming call
When notifying the framework of an incoming call, a lock is held
in both the MmTelFeature and MmTelFeatureConnection. This can cause
a deadlock if the MmTelFeatureConnection is also processing an
event, such as sending/acknowleging an SMS message.

Remove the lock around the listener methods to the framework, since
it is not needed.

Bug: 155083563
Fixes: 155819733
Test: atest CtsTelephonyTestCases:ImsServiceTest FrameworksTelephonyTests
Change-Id: I39f7192c6f79e215ef989797870f5f501197cd08
Merged-In: I39f7192c6f79e215ef989797870f5f501197cd08