Fix logging for successive screenshots
When two screenshots are taken in quick succession, we don't show
UI for the first one (but do still continue the task for saving
the screenshot in the background). However, the logging was done
as part of the UI flow, meaning that we currently don't log when
the first screenshot is saved.

This change fixes the logic flow so that we log success/failure of
the saving task in all cases, but only show UI for the last
successive screenshot.

Bug: 157244123
Fix: 157244123
Test: manual; tested that we don't see the log before the fix and
do after

Change-Id: I7646f4a3cb6cfd0f903472e101fb385cef58ddea