Move pullers to binder thread
Pass in a direct executor to setPullAtomCallback. This will cause the
puller to execute directly on the binder thread, as opposed to the background
thread, which could be busy with other work.

The move to the binder thread means that pullers could execute
concurrently, but these pullers are already properly synchronized, so no
further work is needed.

Test: m
Bug: 157186182
Change-Id: I17d67af49b60c59c7de61aba2ce6716aade6e737