Remove then add during replace to reset vis
This change adds a special case to remove a package from the apps filter
just before adding it during a replace action to ensure we wipe any
no-longer-valid visibility grants. This was broken when we moved the
removePackage call to account for packages that are removed with
KEEP_DATA set. To account for both scenarios, we added make an
additional call for the replace case.

Fixes: 159732800
Test: atest AppEnumerationTests
Change-Id: I031c634986b0666163f0292070be31c33461b9bc