Add null check for resolveInfo and Automotive feature.
AAOS uses multi-user model, and apps become queriable after user is
unlocked. Thus we need to make sure that the querying happens at the
right time with the right user with resolveActivityAsUser.

In order not to affect the stability, we can use the null check +
Automotive check for R.

This null check + Automotive check can be removed when resolving
Activity as User is implemented in S+.

Bug: 158508455
Test: Manual
Change-Id: I58e37f795136508ef7bb53c40214ed3c76f588bf