1/n: Slight changes to ClientMonitor's Callback
1) Rename ClientMonitor.FinishCallback to ClientMonitor.Callback
2) Adds onClientStarted to ClientMonitor.Callback

This will allow owners (e.g. Face10, Fingerprint21) where needed to
keep track of more operation lifecycle. For example, as part of the
resetLockout cleanup that is happening, Face10 (which only supports
a single in-flight challenge) will need to keep track if a challenge
is currently "open", and notify challenge owners if it is interrupted
by another generateChallenge/revokeChallenge from potentially a
different owner.

Bug: 145978626
Test: Builds

Change-Id: Ie51b4e367a07401c604d64a146e881053b7cc4a7