Correct the return value of applyAnimationLocked.
The starting window cannot be removed because it is waiting for the
exiting animation finished, This may happen if the animation is not
applied when #selectAnimation cannot find an appropriate animation.
To make the state correct, the return result of #applyAnimationLocked
should not only reference to the hierarchial animation state.

Bug: 161847146
Test: atest WmTests
Change-Id: I69bcd29f2587a3644485a0db05e622f6775cc82f