Invalidate resized bounds on aspect ratio changes
When the aspect ratio is changed on the pinned window,
the previously saved user-resize bounds are used as
the restore bounds when unexpanding, which have the previous
aspect ratio still.
This change invalidates the user resized bounds on aspect
ratio changes, which then forces it to be re-set to the new
normal bounds.

Bug: 163420655
Test: PIP portrait Youtube video, touch to expand menu, click
'next' icon to switch to a landscape video. The landscape video
should restore to its unexpanded bounds with the correct aspect
ratio.

Change-Id: I808993a921e3dc3858fad2c131e6a86bf45955e0