BatteryStatsService StatsLog.write calls out of lock
BatteryStatsService noteStart() calls are now all processed
on a separate thread. StatsLog.write calls inside of this class
(for convenience), however, should not be processed on that thread -
they should be processed immediately to ensure that their timestamp
is accurate. Since statsd already processes these calls on a separate
thread, there should be no concern. Moreover, these StatsLog lines
do not need to even be in the same mLock.

Test: compiles
Change-Id: Ic3ee78dfffe367bf136de16f39af7371ddf71edd