Add Tap falsing to the FalsingManager.
This moves some of the functionality of DoubleTapHelper into the
FalsingManager. This also gives the FalsingManager an opportunity to
approve single-taps.

Currently, all single-taps are rejected, but future work may enable
single-tap falsing.

DoubleTapHelper is renamed to NotificationTapHelper to better reflect
its purpose.

The FalsingManager now maintains a record of past gestures for
analysis, wherease prior it only ever kept the most recent tap.

Bug: 172655697
Test: manual && atest SystemUITests
Change-Id: I58f3d25881b0a3ce14e47b72287dd6a1a9307852