Fix IME layer
Before, when the app is animating, we have the IME window parent to
assign layer based on the order of the window children. However, the IME
surface parent can be different from the window parent. In such case,
the layer can be wrong.

Before ag/12218308, the IME window parent DA contains IME container and
TDA, so the layer it assigned to IME is always 1. Now, IME window parent
has only one child, so the layer it assigned is always 0. It was a false
positive scenario before.

Fix: 173618785
Bug: 156355859
Test: manual: reproduce steps in bug
Test: atest WmTests:ZOrderingTests
Change-Id: I6f5d741acbd298a22dcbf92336b423579fd1678e