Move externally synced framework files to an 'external' folder.
Java package paths and the overall API is unaffected.

This makes it much easier to maintain the export script. Prior to this
CL, the export script had an allowlist of framework files to leave
alone. This list was badly out of date and hard to maintain in
lockstep with framework.

There are no functional changes in this CL.

Bug: 162450968
Test: m -j, framework-appsearch.stubs.source-update-current-api, presubmit
Change-Id: I5a91f4afb1cfe2aa5e514d5f3159bf212f7de03b