Restore overridePendingTransition to a sync binder call
Otherwise the calling activity is not blocked and can proceed
to change activity state, that may cause the override request
to be dropped. In addition, it is usually called after
startActivity and finish which are non-one-way binder calls.
The order also needs to be guaranteed to make the override
take effect.

Bug: 161781274
Bug: 179008865
Test: WindowUntrustedTouchTest# \
      testWhenExitAnimationAboveThresholdFromSameUid_allowsTouch
Change-Id: I7ca018a934972c195040f2a9f63a18a59dce9c48