Added options to not stop previous users on switch.
On automotive (or on any device that sets
config_multiuserDelayUserDataLocking to true), the current foreground
user and its profiles are stopped when the system switches to a
different user, which causes CTS tests that switch users to fail (as
the test process dies when the user is switched).

This CL addes a new system property (fw.stop_bg_users_on_switch) that
allow tests to override this behavior. It also:

- Splits mDelayUserDataLocking logic into 2 variables (the other is
  mStopBackgroundUsersOnSwitch, which is the one that can be changed)
- Adds more optional logging.

Test: adb shell setprop fw.stop_bg_users_on_switch 0 && \
adb shell dumpsys activity users | grep shouldStopBackgroundUsersOnSwitch
Test: atest android.car.apitest.CarUserManagerTest#testLifecycleListener

Bug: 180515783
Bug: 180515783

Change-Id: I02bbe7b502c94f02a94965b8ff25915b6b55ae1b