Use padding in pages instead of pageMargin
ViewPager pageMargin does some rounding when calculating the offset of
each page, leading to positions that are not integers at the snapping
points. This is usually not an issue, except that we use those positions
to drive the PageIndicator.

Instead, set padding on the pages and negative margin in
PagedTileLayout. As everything has clip false, this produces the exact
same effect.

Test: manual
Fixes: 187954110
Change-Id: Ie1055a34c26214b012e96e3ec032ce4db576694d