Turn off HBM when AcquiredInfo#GOOD
Prevent unnecessary use of HBM. This change decouples illumination
state from touch state. onFingerUp should always be called whenever
the finger leaves the HBM area, and not only when illumination was
requested (isFingerDown was actually isIlluminationRequested()).
Moves the check to onFingerUp to stop illumination only if illumination
was actually happening.

Bug: 187460696
Test: atest UdfpsControllerTest

Change-Id: I2eebe32990761d2fbf86cab228e4a7221a209b42