Add hidden TYPE_ACTIVITY_STARTED to allow notify activity start event.
The current launcher implementation, the resumed Activity doesn't
enter paused state when entering overview. The launcher will have
a start activity event when entering overview but we don't provide
start event now.

We are not allowed to add new APIs now, we add new hidden constant
constants in ActivityEvent and send the event when the activity is
started.

Bug: 189968849
Test: manual. Checks if the activity start events are received by
ContentCaptureService
Test: atest ActivityRecordTests

Change-Id: I8e4f8b7ac8f24c66ccad9f7e884cac9e20ce3cc6