Skip applying expanding clip ops in native canvas
This is the behavioral change that precedes removing support for
expanding clip ops. Canvas.java still passes the Region.Op value
through JNI but then ops other than intersect, difference, and
replace log a warning and make no change to the native Canvas
state.

At the moment, the Skia build flags still allow for the deprecated
ops, so replace is used as before. In the future replace will be
emulated and the build flags will be updated to make SkClipOp
distinct from SkRegion::Op.

Test: Added new CTS tests for current API and API <= 27.
  See other CL in the topic.
Bug: b/151725198
Bug: b/143352151
Change-Id: I933a8d18258878761b6010330240ff1e0c678acd