Postpone available check for CustomTile
In the case of custom tiles, we don't know if they are available until
after initialization. Because it runs async to tile creation, the check
in QSTileHost may fail and destroy the tile prematurely. Instead,
postpone it until after we've tried to retrieve the service icon.

Also, add TileService CTS tests to com.android.systemui.qs TEST_MAPPING.

Test: atest SystemUITests
Test: atest CtsAppTestCases
Fixes: 197970498
Change-Id: Id8d4090cfbfaa144b71d78ffd74da8fc1b9e6c6e