Show color layer in task animations when requested
Additionally wrap AnimationAdapter to inject custom onCancelledCallback
which should run when the animation is cancelled.

We can't rely on the onAnimationFinished callback since that doesn't run
when the animation adapter is canceled by the surface animator and it is still
running (i.e. restarted due to a new animation being requested before
the old one finished).

Since the finishedCallbacks are not always called when an animation
adapter is cancelled this cancel callback serves as a way to inject and run custom clean
up code for the animation adapter, which in our case is used to clear
the background color of an animation.

Test: atest FlickerTests:TaskTransitionTest
Bug: 199507257
Change-Id: Ic603b2711c19299f01271131062765255451310a