Updates magnification button when the magnification activation state is changed
When the a11y shortcut or triple-tap gesture is trggerred, it also
change the magnification activation state.
If the activation state is off, the magnification button can be removed.
If the activation state is on, the magnification button can be shown.
And remove unsed callbacks of MagnificationGestureHandler, onShortcutTriggered
and onTripleTapped.

Bug: 201045813
Test: atest MagnificationControllerTest, atest FullScreenMagnificationControllerTest
Change-Id: I2edc918f32ab9a8de02305bff121aedd483f70bb