Make AlertDialog use specific styles per button
As the buttons in an alert dialog are identified with Positive,
Negative, Neutral, use the corresponding style attributes. That way,
they can be styled independently.

Style them independently in SystemUI. This allows for AlertDialog or
SystemUIDialog created with the right theme to be able to use the
default buttons with correct styling.

Test: manual (AddUserDialog)
Bug: 203666386
Change-Id: I4f292193bff5dec90f8dc229a9824d22075d5241