SpriteController: Reparent pointer surfaces under DisplayContent
When using MediaProjection to mirror a display, only surfaces that are
inside the mirrored DisplayContent's hierarcy show up in the mirror
surface. Previously, since pointer icons created by SpriteController
were not part of SF's hierarcy under the DisplayContent's
SurfaceControl, they did not show up in the mirror surface.

In this CL, we reparent pointer sprite surfaces to the the
DisplayContent's SurfaceControl whenever there are surface changes,
ensuring that they show up in screen recordings.

Bug: 202258016
Test: manual: Start screen recording from quick settings tile
Change-Id: Ib404ed36bd13b0fc8cc4a6fb593dc54b3e35dd1a