Ignore folding feature when it is not on the active display.
Ignore the folding feature states that are not reported through
extensions. There is a crash when trying to process a folding
feature while the corresponding display is not active. We had a
fallback value for the posture of flat but the bounds that we
calculated would be incorrect.  We removed the default value
and ignore the folding feature when the corresponding display
is not active.

Bug: 204477995
Test: manual - install the samples app from WindowManager
Test: close a folding device and open the samples app
Test: select the display feature sample.
Test: app should not crash.
Change-Id: I102ed61618915ae6782c40964882a54184538e81