Store displayIdToShowIme in an IMMS field
This change allows the IMMS to remember the last display id
the IME was requested to bind on.

One use case is when we want to prevent binding the IME if
there is no currently focused editText, and yet we want to still
be able to handle an explicit request to show the IME later.
Remembering the display id helps to show it on the correct screen.

Bug: 199887357
Bug: 37617707
Test: make
Change-Id: I98459c4a44416fd7209150932a2a2520612da818