Fixed divider rendering in RTL languages
In horizontal layouts, when elements are rendered from LTR, the
definition of "before" and "after" gets flipped, causing the divider
spacing to be allocated to the wrong locations. By flipping the check in
RTL layouts, the dividers can be placed correctly.

Additionally, it is necessary to add divider measurements to all layout
width measurements to allow for dividers in right gravity to layout
properly.

Bug: 208927871
Test: atest android.widget.cts.LinearLayoutTest
Change-Id: If6e917873571484a6fb1dfc28b356289e2cc1ffa