Load rotation settings when creating a display
In ag/16433357 applyRotationSettingsToDisplayLocked
was extracted from applySettingsToDisplayLocked method.
This was done to load rotation settings only after updating
the display size.
After separating this method we didn't call the second method
in all places where we called the original one.
Added a flag that controls loading rotation settings
which is 'true' by default to fix loading of these
settings on the first boot.

Fixes: 212200866
Test: atest WmTests:DisplayWindowSettingsTests
Test: open portrait only app on a display with
 'ignore orientation request = true' after booting
Change-Id: Id248393dc389b54c140a83d3a0bb93fef1a200ed