Enable notification permission enforcement
This change will guard posting notifications with a new
"android.permission.POST_NOTIFICATIONS" permission introcuded as part of
Android T.

This permission is marked "dangerous" and will require both declaring
usage of the permission in AndroidManifest.xml as well as showing a
runtime prompt to the user in order to be granted the permission.

Packages which are not granted the permission will have their
notifications silently dropped.

Test: atest Cts*
Bug: 194833441

Change-Id: I274738d4d4c3b09902844708327feb2d1d367307