Fix IndexOutOfBoundsException on VirtualDisplay removal
We were accidentally calling remove() on an ArrayList<Integer> using a
primitive integer, which is interpreted as an index, when we intended to
be removing by Integer value. This CL fixes that by switching to using a
ArraySet for the data structure, since we were only using it as a Set
anyway and don't need List semantics.

Test: atest VirtualDeviceManagerServiceTest
Fixes: 215381936
Change-Id: I730f5f4001fd6b3627d2ab70e7af5860ce367049