Add required flag to registerReceiver call in ScreenshotController
Android T adds support to allow a runtime receiver to be registered as
not exported, but to ensure apps can take advantage of this, calls to
registerReceiver must specify a flag indicating whether the receiver
should be exported for apps targeting T+ that are registering for
non-system broadcasts. This commit adds the RECEIVER_NOT_EXPORTED
flag to the receiver in ScreenshotController since this broadcast is
only sent by the local app.

Bug: 161145287
Test: Build
Change-Id: I35997c130063b075c7f79052c7e82cd476aff12e