Uses ActivityLaunchAnimator for occlusion animations.
This is used when the lock screen is visible and an activity occludes it,
such as the camera. Currently, the camera launches behind the lock screen
and then the lock screen janks out of the way to show the camera.

There are currently some pretty severe performance issues, but I'm not sure
yet whether that's due to the new launch animator, or the fact that the
old animation is so janky that you can't really tell. I'm investigating
that, but we should get this into the build to start working out any bugs
with the remote occlude animations.

This requires remote occlude animations to be enabled:
adb shell setprop persist.wm.enable_remote_keyguard_animation 2

Test: atest SystemUITests
Bug: 197312643
Change-Id: Id801e44c25c6e2340b4269d215b9a1c056f0ceff