Calculate freeform bounds for fullscreen task
Currently several pieces of logic in LaunchParamsModifier assume
that bounds doesn't need to be calculated for fullscreen tasks, but
this is incorrect in that the bounds calculated in
LaunchParamsModifier can be used as last non-fullscreen bounds for
fullscreen tasks. Last non-fullscreen bounds is useful in an
environment such as ARC++ where the windowing mode of a task
changes between freeform and fullscreen.

With this CL, freeform bounds calculation (ActivityOption, layout,
getTaskBounds) is applied to fullscreen tasks too.

This CL also involves some refactoring around getting bounds from
ActivityOption and WindowLayout (especially for fullscreen tasks
and tasks whose windowing mode is MULTI_WINDOW).

Bug: 185427982
Test: atest WmTests:TaskLaunchParamsModifierTests
Change-Id: I3ccaf1ebfa7bcc11c371b0043fbfa11366f7a8f3