[13/n] Letterbox Education: change dialog width and handle oveflow cases.
We set the dialog width to the maximum between 472dp (to be consistent with the standard android dialog width) and the task width minus margin.

We also add a 16dp margin between the dialog container and the task
stable bounds (task bound excluding insets).

Overflow in width will be handled by texts wraping to multiple lines and
overflow in height will be handled by making the dialog scrollable.

In addition, a 500ms start delay is added to enter animation.

Bug: 215314668
Test: N/A
Change-Id: I4b11096e124ae8632bff6390700e7519acdcefa0