Prevent optimization of common code between framework/sysui
This is a temporary fix to ensure that SystemUI builds code that is
consistent with common code that might be referneced in the the bootclasspath.
Note that a proper fix would be to entirely eliminate these duplicated classes,
but for now this avoids any related issues, and eliminates issues when
running coverage tests against optimized builds.

Bug: 219615928,222468116
Test: m -j EMMA_INSTRUMENT=true EMMA_INSTRUMENT_FRAMEWORK=true
  + validate successful coverage test runs
Change-Id: I67b7ad1d6f8e1fb2352652d452c620d195d5c0dc