Add window_frames_valid to proto
This is a step to move the window layout to the client side.

Once the window layout is moved to the client side, when something (such
as configuration) is changed, the server will send MSG_RESIZED to the
client. Before the client reports the new frame to the server, the frame
at the server side might not be up-to-date.

This gives WindowManagerState a signal to decide if it should retry.

Bug: 161810301
Test: presubmit
Change-Id: Idece0f559e7013ff389cbb7aa74d63734c95f64c