Don't perform another layout after finishPostLayoutPolicyLw
If the visibility of any insets is changed, InsetsSourceProvider will
post LAYOUT_AND_ASSIGN_WINDOW_LAYERS_IF_NEEDED, and we will perform
layout there. The logic in finishPostLayoutPolicyLw only checks the
visibility change of status bar, which doesn't cover all insets. We rely
on LAYOUT_AND_ASSIGN_WINDOW_LAYERS_IF_NEEDED instead.

This CL also moves the logic about topAppHidesStatusBar to
updateSystemBarsLw, after updating mForceShowSystemBars, because it
affects topAppHidesStatusBar.

This is a step to move the window layout to the client side.

Bug: 161810301
Test: Rotate Camera and see if it can play ROTATION_ANIMATION_SEAMLESS.
      Rotate Camera with notification shade expended and see if it can
      play ROTATION_ANIMATION_ROTATE.
Change-Id: I94e06a99757fa97376d5644865728aa28c1d6883