Fix bubbles IME alignment on tablet
There was a mistake in the math here, instead of using the y location
of the top of the IME, it was just using the height of the IME in
the calculation. This caused the bubbles to move out of the way
of the IME in tablet portrait when they didn't actually have to move.

This also impacted tablet landscape causing the bubbles to NOT move
out of the way of the IME when they should.

Test: manual - have bubbles on tablet in portrait and have the IME
               show, bubbles should only move if the IME would
               overlap with them.
             - have 5 bubbles on tablet in landscape and have the IME
               show, bubbles should move above the IME as much as
               they can without going past the top of the expanded
               view.
Bug: 223912157
Bug: 215063840
Change-Id: Ifd763caef6c521b1083757cade2898b1361198cd