Remove fossilized upgrade code.
* Handling of device owners without component names.
* Migration of user restriction from UserManager to DPMS.
* Handling legacy owners format.
* Legacy tag TAG_DEVICE_INITIALIZER in owners.

This code can only be invoked on a device that is upgrading from
M or below straight into T+, so the device should have a lifespan
of 8 dessert releases. This is more than twice than any device
officially supports.

This change is needed to make further changes for the issue
simpler.

Bug: 218639412
Test: atest com.android.server.devicepolicy.OwnersTest
Test: atest com.android.server.devicepolicy.DevicePolicyManagerServiceMigrationTest
Change-Id: I142ac7f690ebb5fb7a94cc32306aeb4a6321e541