Fix notifications trimmed when unfolded on keyguard
When the number of notifications is limited by mMaxDisplayedNotifications (e.g. on keyguard), it is possible to use the intrinsic content height of the stack, as it is already limited by the maximum number of notifications. The values of bottom and top margin previously used in updateStackEndHeight were wrong for the split shade (while on lockscreen).

Fixes: 214504318
Test: Add many notifications and check nothing is truncated both when folded and unfolded.
Change-Id: Iae63c3916499c53475475905383c925b13d8f1b1