Fixing setting mQsExpandImmediate state in correct context in split shade
Currently mQsExpandImmediate was set too often - always when panel was tracking touch and sometimes in times when we don't want it e.g. swiping up and/or in locked shade.
Now this code is moved into more scoped callback, when it's called only when shade is actually expanding and only in unlocked state.

Fixes: 230340979
Test: manual + unit tests
Change-Id: I47870107b339b9cde0c61478c68f2f118ff97d62