Add FreeformTaskTransitionHandler
This change ties the window decoration lifecycle to shell transitions if
it's enabled, so that we can leverage the new transition handler to hand
over window decorations.

This change exposes the transfers of management of tasks to task
listeners to implement maximizing. Eventually I want to use it to
implement the passover of window decorations across listeners.

This change lets freeform code handle maximizing transitions because
there isn't anything special to do for fullscreen tasks with window
decorations so we can keep FullscreenTaskListener simple.

Currently all animations related to freeform tasks are jumpcuts because
there isn't any spec defined.

Bug: 205580520
Test: Open, close, maximize freeform tasks with the shell transition and
the caption in shell enabled.
Test: atest WindowDecorationTests

Change-Id: Ie7bb92e0b5946b006fa685f4438815627c72dffe