Do not update visibility inside clearing task
Otherwise if the top activity doesn't occlude and there is a
pending global config change, the activity behind will be
scheduled a relaunch item and a destroy item. Then there may
be unexpected behavior by the onCreate from the relaunch.

Note that isClearingToReuseTask() is only true when starting a
new activity with clearing task. The visibility will still be
updated later when resuming next activity.

Bug: 238978283
Test: atest TaskTests#testPerformClearTop
Change-Id: I4b926adfef0a159f99d002f78838654e9f27baee