Ignore notification update for not active user
We can receive notification updates for the not active user. Ensure that
we do not pop up a bubble in these case.
Store the notification key for these notifications in saved bubble data
for the user that should get this bubble. When changing user back to the
user that should get the notification, bubble will get restored as well.

Bug: 237314340
Test: atest BubblesTest
Test: Enable bubbles for incoming SMS in Messages app. Switch to
secondary user. Get incoming messages and observe that they do not
result in a bubble. Switch back to primary user and observe that bubbles
are shown for SMS that came in.

Change-Id: Ifcabb8458c72edc06f8cff844ac5e2266f5387f9