Adding an API to cancel all alarms
In some cases, an app can lose access to the pending intents it used to
set previous alarms. In the current state, such alarms can never be
cancelled. This can be an issue because there is now a maximum limit to
the number of concurrent alarms an app can set.
An API to cancel all pending alarms set by the caller should help
mitigate this.
It should be noted that an API to query or retrieve all pending alarms
is not straightforward with the current design.

Test: atest CtsAlarmManagerTestCases:BasicApiTests
Test: atest FrameworksMockingServicesTests:AlarmManagerServiceTest

Bug: 182859584
Change-Id: If8f81f93838bf366887c63ae006f9291ee636f5c