Fix the SetConnectivityStatus callback
SetConnectivityStatus callback should only update the InternetTile when
there is no default network because this callback doesn't contain the
updated the information of the WiFi or Cellular.

When there is no default network, InternetTile should show the globe
icon so SetConnectivityStatus should trigger the tile refresh.

When there is default network, InternetTile should be updated by the
corresponding callback (setMobileDataIndicators or setWifiIndicators),
while SetConnectivityStatus is supposed to only update the
mNoDefaultNetwork field.

This CL also added some QSlogs specifically for the InternetTile.

Bug: 241655191
Test: Unit test, manual test
Change-Id: I6560e016ee1663779b3c59d45520ef98507cad77