Fixes bug where home controls would show up before first unlock.
Based on the attached bug, there is an issue where the quick affordance
for home controls shows up on the lock-screen before the lock-screen is
unlocked for the first time. This is a problem that only exists with the
new implementation of the refactord KeyguardBottomAreaView and not in
the original code.

The fix is to take the visibility from ControlsComponent.getVisibility
into account. This CL does that. This is what the old implementation is
doing.

Bug: 244296596
Test: CL was written in a test-driven development (TDD) format. The
failing test was written first, then the production code was written and
that failing test passed. In addition to that, manually verified that
(1) the home controls affordance no longer appears on the lock-screren
after a reboot and before unlocking for the first time and (2) it does
appear normally and behaves normally when clicked after unlocking and
locking again.

Change-Id: I49ef150514cd46d387a94a499d56c18ef2c7e0ba