Fix QS squishiness on LS shade transition
QS tiles much smaller when pulling down shade on LS compared to same
gesture on Launcher. This is because the squishiness fraction is set
differently for LS and Launcher. So, the solution is to make sure the LS
squishiness fraction is in the same range as the Launcher's, so the QS
tiles are the same size on both interfaces.

Test: manual and atest qsSquishTransitionFraction_noStartDelay_startFractionSet_returnsBasedOnStartAndDistance, qsSquishTransitionFraction_startDelay_startFractionSet_basedOnStartAndDistanceAndDelay
Fixes: 243258790
Change-Id: I20f2a9829a972668021a3cff2b2673f2b6967928