Switching UserSwitcherController implementation.
Introduces a class that implements UserSwitcherController by checking
the feature flags and delegating to the old implementation if the flag
is not set.

Right now, the class will throw an exception if the flag is set to on
because all of its methods are unimplemented. In the next CL(s), we will
be implementing all of these.

Bug: 246631653
Test: Ran system UI, did some user switching, verified no crashes.
Change-Id: Ia5df853c5297ba6a086db867ad39111ff6758b6d