Add support for rounded notification in groups
- Added `NotificationTargetsHelper` helps to find viewBefore/viewAfter for notifications inside of a group
- Added `Roundable` interface could be applied to Views and ViewWrappers

Interface that allows to request/retrieve top and bottom roundness (a value between 0f and 1f).

To request a roundness value, an `SourceType` must be specified.
In case more origins require different roundness, for the same property, the maximum value will always be chosen.

It also returns the current radius for all corners (`updatedRadii`).

Test: atest NotificationTargetsHelperTest
Test: Manual tests (phone and large screens)
Bug: 184682172
Change-Id: I39b96f79a00fa358b80e62dbb090f01f60b8d718