Make sure that a recents-launch with remote gets its own transition
If a remote is attached to an activity-launch, it should be its own
transition so that the requester can get the proper transition
lifecycle events (handleRequest). Unfortunately, the current
startActivity API is synchronous, so we don't know if the caller
needs an accurate result or not so normally we have to merge all
launches meaning the remote can get lost.

Fortunately, for recents, we do know that launcher cares about
the transition lifecycle and not the startActivity result. So,
in this case, we can defer the activity start if necessary.

Bug: 250012839
Test: launch app with dialogactivity on top. Swipe to recents.
Change-Id: I715518bf9611b8ad50bbfb5b05dea8002cb5a6e5