Tag UnsupportedAppUsages in SurfaceView with alternatives
There is little evidence that the fields annoatated with
@UnsupportedAppUsage are actually used, and even if they were by some
long-tail of apps there are either public api alternatives to accessing
these fields or there are no good use-cases. Populate the usages with
publicAlternative so that we can start removing them during future
refactors.

The only @UnsupportedAppUsage that remains untouched is
SurfaceView#setFrame, because View#setFrame is also annotated with
@UnsuportedAppUsage and possibly has a higher usage level.

Bug: 249858453
Test: builds
Change-Id: I51c918c10ca087d87791899f7bce4b1425130902