Adds toggle support for quick affordances.
Introduces the ability to toggle quick affordance buttons on the
lock-screen.

Developers of keyguard quick affordances will now have the option of
providing an ON or OFF state for their toggle (or leave it unset) and
the reset of the system will reflect that state onto the UI.

Bug: 253094565
Test: Temporarily added a fake toggle config to the collection of
keyguard quick affordances and made sure it looks correct on both dark
and light themes on the lock-screen and toggles appropriately. Also
expansed existing unit tests to cover for the new state.

Change-Id: Idddb4baeb19c603a6663347a7b4319751f02fd2f