Fix concurrency issue in PropertyInvalidatedCache
Bug: 253063488

Ensure the global lock is held while fetching the list of active
caches.  The global lock is necessary while the list is being fetched;
it is not necessary while each cache in the list is cleared.

Added a new test for PropertyInvalidatedCache.onTrimMemory().  This
test will not catch race conditions but does verify that
onTrimMemory() behaves as expected.

Added an entry to TEST_MAPPING so that the PropertyInvalidatedCache
tests run when the source file is modified.

Test: atest
 * FrameworksCoreTests:PropertyInvalidatedCacheTests

Change-Id: Ib3ef355c88cb272334ccfc16f922c057828749ce