Fix vertical size when using new headers
For the new headers, we shouldn't be looking at quick_qs_offset_header.
Instead, use 48dp for QQS offset.

This is needed because with combined headers, the height of the header is always the same. Without this CL, in some devices, there's a lot of blank space under the header, because the system expects them to be larger.

Test: manual, multiple devices and screen sizes
Bug: 215584502
Change-Id: Ic6f4c45eb4e21891a073f11bb328497a7807edac