Add a system API to get the main user.
This adds a system API to get the designated main user on the device.

The documentation for the existing isMainUser API is updated to make
clear that a main user might not exist on the device, particular for
non-phone/tablet form factors, but there will always be at most one
main user.

The CTS tests for isMainUser are updated to reflect that documentation,
and to no longer enforce that the main user is the first user on the
device

Bug: 259476886
Bug: 256624031
Test: atest UserManagerTest
Change-Id: Id1739b188d7c1c0f9006bc473de35a88d09cad95