Add test app SimpleTestIme to test InputMethodService
Add a test app SimpleTestIme under services/tests/InputMethodSystemServerTests/test-apps/ which provides the basic functionality of a
keyboard. Using it as the target package of instrumentation test, which
means we have access to all the class instances running in it under the
tests, including InputMethodService. In this way we can get a real InputMethodService instance, then manipulate and test the states of it.

Add simple test cases to verify the behavior of some public methods like:
  * onStartInput()/onFinishInput()
  * onStartInputView()/onFinishInputView()
  * requestShowSelf()/requestHideSelf()
  * onEvaluateFullscreenMode()/isFullScreenMode()

Bug: 240359838
Test: com.android.inputmethodservice.InputMethodServiceTest

Change-Id: I84f1a1b25eef4a757d8da27807285e08417458b4