BinaryTransparencyService: Schedule measurement job on demand.
Initially, a binary measurement job is scheduled in a recurring
24h cadence.

We are now changing the scheduling of the measurement job
from one that is time based to one that is on-demand, based
on specific events and capping the measurement to be done at
most once within a 24h time window. Qualifying events include:
1. the update of a preloaded app
2. the update of an APEX

However, this change only cover event #1.

Bug: 259590636
Test: Manual.
	1. Manually update a preloaded app
	   - logcat: measurement was attempted, but not rescheduled
	2. Manually install a new app
	   - nothing is triggered
	3. Manuall update a new app
	   - nothing is triggered
Change-Id: I11ecf3588f263638a074f737487758b37ee27f5a