Add nullability annotation for  VISS#onShow()
The VISS#onShow Bundle argument is provided by showSession. The
VIS#showSession() doesn’t do null check, so it is possible the Bundle
parameter in VISS#onShow will be null. The argument has been
@Nullable for a few releases, but we never marked nullability. We add
nullability for the Bundle parameter to avoid the developer skipping
to do null check that may break the application. In Android 14, the
389401c70ad02e79acdc959f0d3a6f8b59f5083c add an id into Bundle, the
Bundle will not be null starting from Android 14 but the argument
was still Nullable in older platforms. We also update the javadoc
to mention this case.

Bug: 178775773
Test: build pass
Change-Id: I175db4e61a2cf5b49b122f94f72c57bb5a154d85