Pipe data from onWakeupEvent to batterystats
The nflog prefix now also encodes long denoting the network handle
which can be used to retrieve the Network object for the network
the packet arrived on. This is required by batterystats to attribute
the network transport this packet came over.

Piping all the relevent details to battery stats for relevant packet
wakeup diagnostics.

Test: Builds, boots. Manually triggering a wifi wakeup emits the
expected log from batterystats.

Test: atest FrameworksNetTests

Bug: 265742148
Change-Id: Ic08ac2a0d7739f406f2c8dcf75ef58484c823b9f