Fixing mQsVisible state in split shade
mQsVisible/qsVisible was always true in split shade and that's not expected by its consumers.
It also means that part of the code was dead path for several months so I deleted that.

The main problem was calculateQsBottomPosition always returning max value.
This function looks like the value changes linearly but for split shade qsExpansion is either 0 or 1. And for 0, it still returned fully expanded QS value.

Fixes: 267190489
Test: testing different split shade transitions: unlocked->shade, keyguard->shade, locked shade -> bouncer, keyguard -> bouncer etc
Change-Id: Id06bb2b92e6d1648236ab45fa0d4de662ebf219c