Backup and restore - Fix non-system users
Use a flat format, rather than a directory structure, to store files
for non-primary users. Shared preferences does not explicitly support
directories. The code was relying on a soon to be removed API, as well
as it being unsupported by the backup helper.

Files will remain stored under "shared_prefs" and "files", but the
format will be __USER_{id}_{fileName}

Test: atest UserFileManagerImplTest AuxiliaryPersistenceWrapperTest
DeletionJobServiceTest
Test: manual run Backup and Restore from Settings
Fixes: 261168533
Fixes: 266918967

Change-Id: Ib9804a2ec46fc2f5d53995c50adc36ff712fc1c1