Replace InputManager.getInstance() with getSystemService() from context
InputManager.getInstance(Context) method was added to create/obtain
the InputManager as part of b/266013036.
For backwards compatibility with the getInstance() method
(without the Context parameter) was then deprecated,
this made the Context member in InputManager nullable.
Here, all usages of InputManager.getInstance() method have been updated
to use the getSystemService() method of the context parameter,
wherever the context parameter is available.

Bug: b/266709559
Test: Replaced one method with another. Existing unit tests still pass.

Change-Id: Iecf0022cb014d57d651615295c16aba082aa7e60