Initialize DirectA11yConnection pid in constructor, not static.
It is possible for the static pid variable to be initialized before the
app process is forked from Zygote, leading to the pid here differing
from the pid used by app-process queries. This causes deadlock as the
app-process query waits on itself instead of using the same-process
message handling logic that returns immediately.

Fix this by initializing the pid per instance of this class,
in the constructor.

Bug: 269765481
Test: atest AccessibilityEndToEndTest (existing tests)
Test: Run the ATF test from the attached bug
Change-Id: I83b7a6330aaf39c5994eebb4b029314d61c10544