Set mScreenBitmap before saveScreenshotAndToast
Currently we don't set mScreenBitmap until after the fork to just
show a toast if user setup is incomplete. This causes us to attempt
to save a null bitmap. This change moves the line earlier so that
we successfully save the screenshot (and show a toast) before user
setup is complete.

Bug: 271217041
Fix: 271217041
Test: manual (on phone flashed to master)
Change-Id: I5f2312acc318d385420532476b5964e344c4b68e