Hide windowing controls in proto1 caption menu
Proto1 is a system wide mode, so entering/exiting desktop mode should
be limited to the quick settings tile instead of adding windowing
controls to each individual task.

Also removes the logic to hide the desktop button in proto1 since
that will have no effect now that its parent is the one being hidden.

Bug: 272059013
Test: enter desktop mode in proto1 and proto2. Only in proto2 should
the windowing controls be visible.

Change-Id: Ica47a8ee14089c11e711c67abbcec46e2fb56fa0