Transitions: Limit how to LOCKSCREEN can be interrupted
When going from AOD, DOZING -> LOCKSCREEN, it is possible for
isDreaming flag to flip to true for a moment, which can cause a brief
pause in the animation. If the transition is actively running, don't
allow it to be interrupted.

Also fix the flakiness from scenario tests and use a more reliable way
to send transitions

Fixes: 265303901
Fixes: 274799450
Test: atest KeyguardTransitionScenariosTest
Change-Id: Ia8d5dfc8e0d2abeeea1c34b6915a8eb6a3d9bc2c