Deprecate PointerIcon#TYPE_DEFAULT API
The PointerIcon#TYPE_DEFAULT constant has the same value as TYPE_ARROW,
which makes it a redundant API. Especially now that styluses can also
show pointer icons, this API is misleading, because not all pointer
types have the same default icon. While the arrow might be the default
for a mouse or trackpad, it will likely not be the default for styluses.
Since this default is no longer correct, deprecate the API.

The way to use the system's default icon is to return a null value for
PointerIcon through the View APIs. This is not to be confused with
TYPE_NULL, which shows no icon.

Bug: 215436642
Test: Build
Change-Id: I89f55685b588dc39a0d2b34619e4187d9f041a56