Disconnect existing adb clients when adb grants are revoked
When revoking previously authorized adb grants, the prompt indicates
access will be revoked for all computers that were previously
authorized; the intent of this was to remove any grants previously
authorized with the 'Always allow' option, but this could also be
read to indicate existing connections are terminated as well. This
commit disconnects all existing sessions by toggling the adbd
service; the original behavior can be maintained for lab
environments by modifying the global Setting
adb_disconnect_sessions_on_revoke to 0.

Bug: 202836162
Bug: 261184503
Bug: 262244393
Test: Manually verfied existing USB and Wifi sessions were terminated
Change-Id: I22266053cf1f8a0452bfa22b2898b4c68eeaed1d