Introduce Client Transient flag (1/n)
This is a feature flag to develop transient related features on the
system UI side. Currently it's handled by the system server.

This change introduced the flag and set it by default to false, in the
meanwhile, wrapped existing server side logic with the flag to make the
future clean-up easier.

If the flag is flipped without system UI side feature complete, the
transient animation will not work. Other insets related functionality
will not be affected.

SystemGesturesPointerEventListener should be simplified or changed to a
class instead of an interface with the feature development.

Bug: 277290737
Test: DisplayPolicyTests
Test: DisplayPolicyInsetsTests
Change-Id: I243407d233d71b40807d29660090e674b8dd5d31