Rename "Absolute Volume Control" to "absolute volume behavior"
Currently, the CEC framework uses "Absolute Volume Control" to refer to
the state where both of the following are true:
1. The System Audio device supports the <Set Audio Volume Level> message
   (that is, it supports the CEC "Absolute Volume Control" feature)
2. The HDMI audio output device in AudioService uses "absolute volume
   behavior"

This state is currently named for (1); this CL changes it to be named
for (2).

This is because we plan to reuse the code path for enabling this state
for a new state, where (1) is not true: the System Audio device need
not support <Set Audio Volume Level>/"Absolute Volume Control".

In this new case, (2) will still be true, as  the audio output device
will use a variant of absolute volume behavior - adjust-only absolute
volume behavior.

Bug: 240663266
Test: make

Change-Id: I679d5c1e74a1bfd8dcfe9246410ccb3089a5c5ac