Camera: Use characteristic from user app for extension
For the advanced extender, camera characteristics are provided from the
extension proxy service. However, the vendor key-values are ripped off
from the characteristics because the service does not hold the camera
permission.

To maintain consistency across the basic and advanced extender and
CameraX extension and Camera2 extension, this change will populate the
camera characteristics from the user app side.

Bug: 279854053
Test: Manual test with vendor key, Camera Extension CTS
Change-Id: Ic08751b824ea1cfa0208cf7246ab8c3ff7cd367c