[SettingsProvider] skip caching in system server
The old implementation avoids closing the memory int array if the
process is the system process. However, this doesn't prevent the object
to be garbage collected. As a result, the array may be left open without
being properly closed.

This change skips sending the memory int array and index if the query
is from inside system server; should also addresses StrictMode complaints.

BUG: 277448672
Test: boots
Change-Id: I309ad3558a77d8e7d0c8a8d2c8f45db5a316c976