Support ghe with upturn on ANSI keyboards using Ukrainian layout
It has been brought to our attention that is is not possible to
use a letter on ANSI keyboards; this is because those keyboards
have one less key than ISO ones.

In Windows, AltGr+U can be used to enter that particular letter,
but in Android, Ukrainian was using the same QWERTY overlay that
is used on Cyrillic layouts.

This CL removes the QWERTY keys that were available with AltGr. I
left the punctuation and number ones because their usage will be
sporadic, but when entering latin alphabet words, we want to
incentivate the user to switch layouts.

Bug: 279207810
Test: Flashed on device, selected Ukrainian layout, and tested
with a physical keyboard

Change-Id: I3fc4652d7c78d9185bf9611800ee5cb1e08d55df