Setting flag for auto confirm feature to always return true
This CL aims at making changes such that auto confirm feature flag
always return true. This would work because we are anyways setting
default value of flag to true. This just removes check to go via the
Mendel route to identify the default value of flag since that can have
some underlying issues.

Bug: b/282246842
Test: Tested by building and flashing on local and checking if feature
is available or not

Change-Id: Ifff9008b2bb4a632131e53959e979b8b57d922b3