Handle split to pip if open another fullscreen app case
When this case, because the pip task is outside split root so we
will start dismiss transition and bring another side stage to top.
It cause transition weird and also caused the target opening task
not on top due to we try to bring another side task to top.

Fix this by add another specific logic to handle this. In this case
, we should dismiss split but do not bring any side to top.

Bug: 282894249
Test: manual
Test: pass existing tests
Change-Id: I8855473f60b58290ad6eac1a005f719b2a5125d4