chore(#MagSettingsPanel): make service notify sysui the controlling magnifier scale changed
Add onUserMagnificationScaleChanged method in IWindowMagnificationConnection. When MagnificationController was notified the controlling magnifier scale changed, it would delegate the new scale to WindowMagnification through connection. Once the WindowMagnification was notified the scale changed, it would cache the new scale for the given user/display. The cached scale would be passed to the WindowMagnificationSettings to update the panel seekbar in next cl.

Bug: 286176069
Test: manually
      atest IWindowMagnificationConnectionTest
      atest MagnificationControllerTest
      atest WindowMagnificationConnectionWrapperTest
      atest WindowMagnificationManagerTest
Change-Id: Ia1402c90d6251f85601f41738c8c59f70a26997e