Add SplitScreenUtils#enterSplitViaIntent for FlickerTests
As SplitScreenUtils#enterSplit uses UiAutomation to enter split-screen
with UI operations, which may flaky if the UI component or position
changed, or the test doesn't requires to verify entering split-screen
with drag-and-drop or clicking the menu on the overview.
(e.g. UnlockKeyguardToSplitScreen)

Added the new helper method to enter split-screen mode with using
EXTRA_LAUNCH_ADJACENT for SplitScreenPrimaryActivity to launch
SplitScreenSecondaryActivity in adjacent side.

Fix: 284175072
Test: atest WMShellFlickerTests:UnlockKeyguardToSplitScreen \
         --rerun-until-failure 10
Change-Id: Ic34007b0fe922bec0cef0a7b2a1bfd80cbea5776