Add animated chip in statusbar on external display connected
This adds an animated chip appearing for ~3s on the top right corner
when an external display is connected, in a similar way as the battery
charging animation.

The battery event arrives much earlier than the display becomes
available. For this reason, when a display is connected first the
battery chip is shown, then it is animated away and after ~1s the
connected display icon is animated in.

Test: SystemEventCoordinatorTest
Bug: 289883166
Change-Id: Ia689a6519fad70ae3157d1da2a69473be701c38f