Fix overlapping Manage and Clear all buttons when display scale is at maximum
Fix the overlapping "Manage" and "Clear all" buttons in the Notification
Shade, when "Settings" > "Display size and text" > "Font size" and
"Display size" are maximized by using a Constraint Layout chain.

Test: Manual
1. Toggle "Settings" > "Display size and text" > "Font size",
and "Display size" to max.
2. Go to Settings > System > Languages > App languages.
Set the app language for Calculator and Calendar to Traditional Chinese.
3. Open the Shade and drag down, the "Manage" and "Clear all" buttons
should not overlap (though they may be next to each other)

Bug: 273401782
Bug: 287893062
Change-Id: Ifac1914f1f5428fc4b06dc141ade66eac1432985