[pm] do not stop system apps with disabled a launcher activity
If an app has a disabled/unexported launcher activity, it is effectively
the same as not having a launcher activity. We should not stop such
system apps by default.

BUG: 286459841
Test: manual with a system app that has disabled launcher activity and
verify that it is in the `stopped=false` state after first boot
Test: also tested updating such a system app and verified that it
remains in the un-stopped state after the update is uninstalled

Change-Id: Ie865aba7325e3442c6610ef962fce3a3a65f0b70