Use @UsesReflection for LocalTransport keep rules
Prefer the @UsesReflection annotation to generate the keep rules
necessary to preserve LocalTransport derived class names. These
class names should be kept as they are used to generate a name that
may be referenced by other packages. Other classes currently in the
proguard config are automatically kept by manifest-generated keep
rules, or are otherwise unnecessary.

This change is effectively a no-op under the current shrinking
configuration for both LocalTransport and EncryptedLocalTransport. If
obfuscation is enabled, this annotation will ensure continued
correctness.

Bug: 248580093
Test: m + ensure all classes kept
Change-Id: I867693861e5b1ca9765c926d676c0b18b16cc340