Take timestamp before causing ANR in the test
Before this CL, the test failed on a hwasan device with the following
exception:

STACKTRACE:
java.lang.IndexOutOfBoundsException: Index 0 out of bounds for length 0
	at jdk.internal.util.Preconditions.outOfBounds(Preconditions.java:64)
	at jdk.internal.util.Preconditions.outOfBoundsCheckIndex(Preconditions.java:70)
	at jdk.internal.util.Preconditions.checkIndex(Preconditions.java:266)
	at java.util.Objects.checkIndex(Objects.java:359)
	at java.util.ArrayList.get(ArrayList.java:434)
	at com.android.test.input.AnrTest.clickCloseAppOnAnrDialog(AnrTest.kt:119)
	at com.android.test.input.AnrTest.testGestureMonitorAnr_Close(AnrTest.kt:89)

To fix this, stop collecting initial reasons before running the test.
Instead, just take a timestamp.

That said, this failure has not yet been reported by our automated
tools. This might be due to the use of hwasan on the local development
device.

Bug: 288343153
Test: atest com.android.test.input.AnrTest#testGestureMonitorAnr_Close --rerun-until-failure 10
Test: atest InputTests --rerun-until-failure 10
Change-Id: Ia62f7161211392f9c63092ff271b394c4ff76e48