Fix CtsAtomicInstallTestCases flaky
There is a timing issue when the session re-opens again after the
session is abandoned. Sometimes test cases will reopen the session
immediately when received callback, but at the time, the session does
not be removed from mSession. That's because the session is really
removed in a runnable that posts after notified callbacks.
To fix the issue, notify callbacks after the session is actually
removed from the mSession. This change doesn't apply to success cases
to avoid adding an unknown delay if somewhere wants to handle install
success.

Bug: 288319099
Test: atest CtsAtomicInstallTestCases --iterations 400
Change-Id: Iec7387eeac2beb19cda788363b6d31fa2399abdb