Add expansion fraction to ShadeInteractor
Mirror NPVC expansion amount in a flow in the ShadeInteractor
instead of a member var. This flow will initially be backed
by a value in ShadeRepository, but it will eventually be replaced
by observing the current scene and scene transition progress,
allowing parts of SysUI to migrate to the ShadeInteractor's notion
of expansion without being broken later on by Flexiglass.

Bug: 280887022
Test: Updated and ran affected tests
Test: Manually verified go/shade-cujs
Change-Id: Ib1a87e7b0e57f324c9088f9c526fd882c963e45d