Using performHapticFeedback on BiometricViewBinder
Migration towards a new oneway API to play haptics from the main thread.
A stateflow is introduced from the PromptViewModel to indicate a
HapticFeedbackConstant to be played in the BiomemtricViewBinder. The binder
collects the flow and plays the indicated haptics. CONFIRM and REJECT
constants are used to indicated success and failure of authentication. A
feature flag controls the migration.

Test: atest PromptViewModelTest
Bug: 245528624
Change-Id: I27e4057e0613ed1b24724ddafd85771da5c3019d