Fix that changes in desiredHdrSdrRatio doesn't update brightness
When onHdrInfoChanged is received from SF and the resulting
maxDesiredHdrSdrRatio changes ensure that the hbmChanged
callback is invoked even though the high-level mode did not
otherwise change. Otherwise the display brightness does not
react to the change, and so changes in the desired HDR headroom
does not result in changes to the actual display brightness.

Bug: 234181960
Test: HighBrightnessModeControllerTest#testHdrRespectsChangingDesiredHdrSdrRatio
Change-Id: I41f8062781493a73d4f1752c6c88d38dfad74909