Overscroll animation for SceneTransitionLayout
We've added a default overscroll for the SceneTransitionLayout. If the
scene receives a speed other than 0 and there is a subsequent scene in
that direction, an animation will be performed based on the speed,
towards the next scene and then returning to the current scene.

To block this default behavior, it is sufficient to consume the
remaining speed in the onPostFling() of one of the children.

Test: mp SystemUIComposeGallery
Bug: 291025415
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:aa3e4bbb1ca4e5b0ef8a54f68b3918107f09c7e4)
Change-Id: Ib0ed74f4a21304bfdc99c2efc619529bf5dc3916