Unify the default behaviour path for solid color ss
ActivityRecord#shouldUseSolidColorSplashScreen is used to force a
solid color splash screen when necessary, taking into account the style
set in the ActivityOptions and the preference set by the activity theme.
When there is no style set in the ActivityOptions and the app has not
specifed a preference, the method provides a default behavior.

Currently this default behaviour is different depending on whether
the ActivityOptions object is null. This shouldn't be the case - the
default behaviour should be the same when the object is null and when the
splash screen style in the ActivityOptions is UNDEFINED.

This CL unifies the default path for when the SplashScreen style is not
specified in the ActivityOptions and the activity doesn't have a
preference for showing an icon.

Bug: 296354973
Test: m
Change-Id: Ia27afaf91fc526b4b34efb40b5a43b0f3796c8c6