Fix recents animation tracking for desktop
Make sure recents animation triggered from desktop can be merged with
new app launch animation.
When live tile is active in recents, the recents animation continues to
run. When tapping on another app to launch it, the launch animation gets
merged with the existing recents animation.
When setting up the recents animation for desktop, we were not correctly
setting the sub animation tracker which caused the merge to be skipped
(it was assumed there are no pending animations).
Skipping the merge caused the recents animation (triggered from desktop)
to remain in a pending state and we never cleared it up. Which left
launcher in a bad state and caused issues for new app launches.

Bug: 297590571
Test: manual, open an app on desktop, swipe up to overview, tap on
      another fullscreen app, check that is launches correctly
Change-Id: I2e1f241b42e55eb0a7b5c3d0b9c956fa036767f5