Persist bugreport mapping in a file
Creates a data/system/bugreport-mapping.xml file, which persists
the (uid, package) to bugreport file mapping. It also stores the list
of bugreport files which should be kept in the framework after being
retrieved by the caller. This file is not on the critical path of
starting the system service. It will be lazily read once per boot at
the first time a caller requests to retrieve a bugreport.

When a new bugreport has been generated with deferred consent, the
mapping file will be recreated.

BYPASS_INCLUSIVE_LANGUAGE_REASON= Existing API name

Test: atest BugreportManagerServiceImplTest
Test: atest CtsRootBugreportTestCases
Bug: 303210021
Change-Id: Ibb7f975a20973f194cf04367d4a24653e9159fb8