Add TEST_MAPPING for time related coretests
Add TEST_MAPPING for android.service.timezone.
Also add TEST_MAPPING for coretests.

For coretests:

There's a TEST_MAPPING at the top level that will pick up tests marked
with @Presubmit, but this is more explicit and runs all tests in a
package when a test has changed. The time coretests should be in good
shape so no filtering here for Presubmit (yet, though could be added as
needed).

Test: treehugger only
Bug: 182461754
Change-Id: Ic537ab0d26adeb300a665eb5b5fe8ea0f123e582