Add new HideInSettingsInQuietMode user property
Defines whether a user should be shown in the Settings app depending on
the quiet mode.
If the property is not set then the user is shown in the Settings app
irrespective of whether the user is in quiet mode or not. If set, then
the user is shown in the Settings app only if the user is not in the
quiet mode.

https://docs.google.com/document/d/1CdjUjAE84-5ZEPRIfw5KYFjLVHtEZxc_sF0w95su8DA/edit?resourcekey=0-dAACT9HRexY1IyoxMmkVlw#heading=h.58jd58rmznte

Bug: 302278487
Test: manual
Test: atest UserManagerServiceUserTypeTest
Change-Id: Iff37f478fe2bb254ee43d7eea1054769938332f1