Use a seperate transaction to applying windowing mode changes after exiting pip
Currently, we always use the wct that is called during
TransitionFinishCallback#onTransitionFinished in order to apply the
windowing mode vhanges after a task exits pip. This method applies the
wct during cleanup rather than applying it in its own transition. This
means that the transition is not sent to the
FreeformTaskTransitionObserver which means, when desktop windowing is
enabled, the view model is not given the chance to create the caption
needed now that the task is out of pip. By using a seperate transition
during non-fixed rotation transitions, this view model will recieve an
onTaskInfoChange call and a caption will be added.

The reason that a separate transition is not used for fixed rotation
transitions is because a new transition will reset the orientation,
positioning, and other states which were set in the original
transition., Given that this change is only required for desktop
windowing which does not utilize fixed rotations when exiting pip, the
wct is not applied as a seperate transition for fixed rotation
transitions.

This change also changes WindowContainer#canStartChangeTransition so
that screenshots are not shown in the cleanup transition as well as the
main pip transition.

Bug: 288910069
Test: Enter pip and then return task to fullscreen. Caption should be
visible.
Flag: EXEMPT bugfix

Change-Id: I2e51880b307f201096f3d40010b61453cc596749

Former-commit-id: d89a020dc77d16fe005c9746cc0f9731ea0dd8c3