Support left & right clipping for BubbleExpandedView
Currently we only allow clipping for the top and bottom of
the expanded view. This CL enables clipping for left and right
sides of the view as well.

This is code is currently unused, but will be used for the new
expanded view animations (which will be initiated behind a flag).

Test: treehugger
Bug: 311450609
Change-Id: I7b00f1352d4de2031e6b744bff8c4a00d594cc01