fix(window magnification): fix onSingleTap_enabled_scaleAnimates presubmit flaky
Refer to ag/25016312 for WindowMagnificationAnimationControllerTest, the test case becomes flaky because the runOnMainSync may take too much time dispatching to the main thread and thus the bounce effect animation is already done. Therefore, we apply AnimatorTestRule in WindowMagnificationControllerTest to eliminate unexpectable animation behavior and the delay of dispatching to mainthread.

Besides, remove MockMagnificationAnimationCallback because we don't use it anymore.

Bug: 305654925
Flag: NA
Test: atest WindowMagnificationControllerTest locally and test on abtd
Change-Id: I8c9d71a6aaf41c0b6cfc6fd6553b8b8e540b8398