Lower EJ execution limit to 10 minutes.
Developers are informed that expedited jobs can be stopped earlier than
regular jobs and so shouldn't use them for long pieces of work.
The standard timeout is 3 minutes for EJs. There's little reason to let
them run longer than the normal 10 minutes, and letting them run up to
30 minutes may give developers the wrong impression and increases the
potential for accidental power issues. Lower the maximum execution limit
of EJs to 10 minutes (the base execution limit of regular jobs) to be
more in line with what EJs should be used for and prevent additional
issues.

Bug: 284512488
Bug: 297106511
Bug: 299329948
Bug: 299346198
Test: atest FrameworksMockingServicesTests:JobSchedulerServiceTest
Change-Id: Ia4950f0801f2881a34976987f735e5ab7f493ee1