Fix an issue where if an app specified a height we wouldn't use it
The bubble API allows apps to specify a custom height for a bubble.
The recent changes to specify a fixed height for bubbles in large
screen landscape didn't account for custom heights.

This CL fixes the issue by returning the fixed height for large
screens in the getMaxExpandedViewHeight method -- that's what gets
called when the developer hasn't specified their own height.

This CL adds some tests for max height / app specified heights.

Additionally this change modifies the height calculation for large
screens slightly to:
- subtract the expanded view padding so that the view has a bit of
  space below the status bar & above the bottom inset
- account for the pointer similar to the other height calculations

Test: manual - go through CtsVerifier test for bubbles
Test: atest BubblePositionerTest
Bug: 310145069
Bug: 310805933
Change-Id: Ib1595894598ab605ee58fc6245d19f32ece58897