FakeGlobalSettings: crash in getContentResolver()
FakeGlobalSettings doesn't actually contain or reference a
ContentResolver, but SettingsProxy has a registerContentObserver method
that callers can use if they're just getting the ContentResolver in
order to register an observer.

Therefore, instead of returning null and letting the caller crash, throw
an exception immediately and include a helpful pointer.

Bug: 261728888
Test: presubmit
Flag: NA
Change-Id: I913711df1c41693157da12fbd4306de1b2d2a7c8