Refactor STL Element maps logic
This CL refactors the logic to create/removes Elements and
Element.TargetValues in the SceneTransitionLayout maps, such that we
don't incorrectly throw an exception when a Modifier.element() is moved
from one node to another in a given scene.

Another major improvement this CL brings is that we don't need
Modifier.element() to run in a composition context anymore, which should
make Modifier.element() much more performant.

Bug: 310241171
Bug: 291566282
Flag: NA
Test: atest ElementTest
Change-Id: I1f513191bca1cbff4917091b9af069c47ea658da