Compute transformations when transitions start
This CL changes when we compute the transformations associated to a
transition.

Before this CL, we assumed that transition specs would never change and
their associated transformations were cached. Because we want user code
to provide different transformations depending on app state (see
b/308961608#comment13 for more context), this CL changes that and
computes the transformations of a transition right when the transition
is started.

This CL also makes SceneTransitionLayoutImpl more leightweight by
removing the need for Snapshot-backed data structure, making sure that
we rely on those only for data that is read during composition or
written during composition and outside effects.

Test: PlatformComposeSceneTransitionLayoutTests
Bug: 315763713
Flag: N/A
Change-Id: Id432d5febdc2d84a9bdaab1a9704957d377756ad