Introduce InputMethodUtils#splitEnabledImeStr()
This is a follow up CL to my previous CL, which added

  InputMethodUtils#getEnabledInputMethodIdsForFiltering() [1]

as an internal utility method.  This CL extracts out its main logic as

  InputMethodUtils#splitEnabledImeStr()

with a unit test so that we can reuse the same functionality in
subsequent CLs.

This is a mechanical cleanup. There must be no observable behavior
change.

 [1]: Iac98fdbb2758f4d9c68930f49d219eb83e54a3d0
      6c984cca294bc5022f3588a04e339c142a7c7c64

Bug: 261723412
Bug: 309837937
Test: atest InputMethodUtilsTest#testSplitEnabledImeStr
Change-Id: I503c0023b62bec949d1681204cc1ad4481675b00