Synchronize window config updates (15/n)
When bundleClientTransactionFlag is enabled, we want to queue all
ClientTransactions before dispatch, but we also need to wait
handleResizingWindows() in order to dispatch WindowStateResizeItem in
the same transaction.

This cause an issue to the timing of other *organizerController pending
transactions dispatch. Move the handleResizingWindows() and
ClientTransaction dispatch to the front.

Bug: 260873529
Bug: 316859294
Test: manually verify with AE demo app.
Change-Id: Iae00a9632dc61d66d976d2d43635190c07a6ce91