Bind silent section clear listener in the viewbinder.
This finally allows us to decouple the notifstats from nsslc, to keep
the source of truth in the repository/interactor.

I also removed the shouldHideParent method because it was making things
harder to follow, since the whole "parent" idea doesn't really fit for
the section header anyway. Plus, it was calling into the controller for
every single child, now we're only calling it for the silent section
(or not at all if the refactor flag is enabled). Didn't add any real
behavior changes, just folded the logic into getVisibleViewsToAnimateAway.

Bug: 293167744
Test: NotificationStackScrollLayoutTest, NotificationListViewModelTest,
ActiveNotificationsInteractorTest
Flag: ACONFIG com.systemui.notifications_footer_view_refactor STAGING

Change-Id: I0224cd0cef4326da50061e39094bbbe57fc49032