Hide time on Tile for alarms further than a week
Also shows month and day on the tile to signal next week(s).
This fixes a bug where pausing an alarm for a week would not get
reflected on a tile since the tile only dispaled day of week and time.

Fixes: 317279939
Flag: aconfig com.android.systemui.qs_new_tiles DEVELOPMENT
Test: atest SystemUiRoboTests:com.android.systemui.qs.tiles.impl.alarm.domain.AlarmTileMapperTest
Test: atest AlarmTileMapperTest
Change-Id: Ic9ccafdaff22a2b745bf91f1454aefc862688586