Animate visual indicators between all types.
Enables visual indicators to animate between any two indicator types.
This is done by the addition of
VisualIndicatorAnimator.animateIndicatorType, which takes two types as
arguments and interpolates between the bounds of each.

For indicator transitions involving INVALID_INDICATOR, we animate by
fading in or fading out based on whether our initial or ending type is
INVALID_INDICATOR respectively.

Additionally this CL merges some of the drag logic in
DesktopTasksController that is largely the same between freeform drags
and drags through status bar.

Still to implement for this bug:
-New drag zones per starting windowing mode
-Tying windowing transitions to indicators.

Bug: 280828642
Test: Manual
Test: atest DesktopTasksControllerTest
Change-Id: I3afcc6fad2e49ae4f4854900cc5819d3a6c254f6