MISSION UPDATE // FLAGS
This is intended as a minimal reference CL for setting up a new flag,
and including a tiny amount of new code behind that flag.

In what is truly a shocking coincidence, the new code plants a little
yellow flag on the planet you've landed on.

Bug: 320150798
Test: adb shell am start -n android/com.android.internal.app.PlatLogoActivity
Flag: systemui:flag_flag
Change-Id: I9527df96a4fb3d0b73704ac2a455bf8820c56951