Adjust padding for communal widgets to be centered vertically
Remove extra padding added on AppWidgetHostView to allow widgets
to occupy the entire box. The adjusted padding is to leave just
suffcient space to render the outline when a widget is highlighted.
The CTA tile now renders in the same height with the widgets.

Bug: b/320770336
Test: manually
Flag: ACONFIG com.android.systemui.communal_hub DEVELOPMENT
Change-Id: I7746f509886edcd5f94968278e4f6f7f558421b9