Replace swipe Edge (detector) by SwipeSource (detector) (1/2)
This CL replaces Edge and EdgeDetector by a more generic SwipeSource and
SwipeSourceDetector. This will allow users to define their own
SwipeSource's, for instance to implement a dual shade where swipe down
from the left half of the screen has a different behavior than swiping
down on the right side of the screen.

Bug: 322130260
Test: FixedSizeEdgeDetectorTest
Test: SwipeToSceneTest
Flag: N/A
Change-Id: Ib50c75ea7bb5a9fa0d775845a2ee16acb4aceecd