Decouple blueprints with preview renderer.
With changes to our layout infrastructure, the preview renderer is
susceptible to regressions. Currently, we are seeing duplicate smart
spaces here.

There is also an exception when we are invoking requireViewById on the
large clock when it is not inflated there. This is because the section
inflates the large clock on a Flow which can not be inflated before we
try to find it.

In order to fix these problems, we are just going to decouple the views
from the blueprints. The views in the renderer are pretty simple and we
can look for a better solution in the future.

Fixes: 319142287
Fixes: 322234032
Test: flexiglass on and off and look at preview renderer.
Flag: NONE

Change-Id: Ie16f2c858a438979851f45c8a3fc897c1c896208