Enable some QS `multivalentTests` under Ravenwood.
This change is a mechanical adding of the `@EnabledOnRavenwood`
annotation to tests that are known to be passing green under the
Ravenwood environment.

From an earlier change, the default for all `SysuiTestCase` is to
be excluded/ignored when running under Ravenwood, except where a
more-specific annotation overrides it.  (Which is what this change
is doing.)

Bug: 319647875
Test: atest SystemUiRoboTests
Test: atest SystemUiRavenTests
Change-Id: I6bfa6376127d14a1edbf7aa995c9b885284c0144