Add SystemServiceRegistry.getService(String).
Allow services to opt in to being fetched entirely without a Context.
This is not enabled by default because it would be a compatibility break
for the service to need a Context in future, but if a service can commit
to never needing a Context, this allows additional flexibility for
callers.

Bug: 319292658
Test: Tested indirectly in followup CL using the API
Change-Id: I4e8a15420a207c93bfa07e41e45edbd06e548204