Move OverviewProxyService from onUserChanging to onUserChanged.
This CL effectively reverts ag/22336171. Which was introduced to
prevent app icons blinking after a user switch, by waiting for the
launcher's main handler thread to become available and indirectly
stalling the dismissal of the user switching dialog. This wasn't a
very good solution but it was hiding jank in favor of performance.

Now this workaround is not necessary, since app icons blinking is not
happening anymore even after this CL. But with this change, duration
of the UserSwitchObservers part of the user switch decreased by 75%
from 971.58ms to 243.48ms and the standart deviation decreased
10 times from 475.40ms to 41.80ms. Comparison charts and more info are
attached to the bug.

Bug: 319352292
Bug: 275349347
Test: Manual control of Perfetto traces
Test: Visual control of user switch
Flag: NONE
Change-Id: Ic78c392ab0f1d0235154953691f7f5dfae19a474