Fix touch handling for glanceable hub over dream
When dreaming, the glanceable hub can be opened on top and touch input
should go to the hub.

This change uses the hub's transition state to force
NotificationShadeWindowView to be visible instead of the
ForcePluginOpen APIs, which seem to affect touch handling.

Bug: 322541769
Fixed: 322541769
Test: atest CommunalInteractorTest NotificationShadeWindowControllerImplTest CommunalTouchHandlerTest
Flag: ACONFIG com.android.systemui.communal_hub STAGING
Change-Id: Ib414f0a503df0cb72c94d3819b58bbbc21ff4fba