Ensure that resetStateLocked is called
when NOT_READY. When we disable the esim, the state will be NOT_READY
(6). We need to ensure that we move away from the sim screen when we
disable the esim.

There is a comment here that says that we do not want to show lockscreen
when SIM is NOT_READY; however looking at the CL, there was a different
ensuing code path that will invoke #doKeyguardLocked. #resetStateLocked
should be fine though.

Fixes: 317762890
Test: disable esim
Flag: NONE
Change-Id: I7f8a0f11be8f3343cd8d0dcc01559642a9dd6509