[flexiglass] Fix incorrect error message positioning bug.
The reason the attached bug occurs is that the Crossfade is always as
wide as the wider of the two currently showing elements and doesn't
center-align them. During the
cross-fade animation, when both the longer "Enter your pattern" and the
shorter "Wrong pattern" messages appear with the shorter being
start-aligned to with the longer, making it appear not center-aligned
within the parent. Once the longer message is removed, the shorter one
becomes center-aligned again.

The approach is to always center-align the messages.

Fix: 325279799
Test: on phone in portrait only.
Flag: ACONFIG com.android.systemui.scene_container DEVELOPMENT
Change-Id: I957168f60a084a27ee7a7aaabc6c1600a716f5d6