Add UDFPS transition support for DOZING
The device entry icon should follow AOD logic
when DOZING, because if the device pulses (from
an incoming notification or from the quick pickup gesture),
we show the AOD UI while the device is dozing. Therefore,
we need to put the device entry icon in the AOD-UI state
when dozing.

Test: atest AlternateBouncerToAodTransitionViewModelTest AlternateBouncerToDozingTransitionViewModelTest DozingToGoneTransitionViewModelTest DozingToLockscreenTransitionViewModelTest DozingToPrimaryBouncerTransitionViewModelTest GoneToDozingTransitionViewModelTest LockscreenToDozingTransitionViewModelTest PrimaryBouncerToDozingTransitionViewModelTest
Test: manually test transitions to and from DOZING
Fixes: 314911640
Flag: ACONFIG com.android.systemui.device_entry_udfps_refactor DEVELOPMENT

Change-Id: I57c0baf88f84adb72e930ebaa24790496cc67be5