fix(magnification): FullScreenMagnificationGestureHandler state decision is ambiguous
In FullScreenMagnificationGestureHandler#DetectingState, when deciding the state transition for MOVE event, there exists some ambiguous logics. If mIsSinglePanningEnabled is true and the event causes overscroll, the handler state will first transit to DelegatingState then transit to PanningState right away, which is not expected. If the condition matches, the expected state should just be the DelegatingState. Threrefore we move the PanningState transition into else block so the transistions will not happen simultaneously.

NO_IFTTT=checked

Bug: 319537921
Flag: NA
Test: manually
      atest FullScreenMagnificationGestureHandlerTest
Change-Id: I0157ddd4117ff4048c3bc44e614cb73c7425dd3f