Skip dialog to unlock private space while launching apps
The UnlaunchableAppActivity handles the prompt to show the dialog to
"unapuse work apps" when a work-profile app activity is accessed but the
work-profile is paused. However, the prompt is shown for all
profiles and not just the work-profile. This results in the prompt being
displayed for private profile with the "unpause work apps" message. With
this change, we skip showing the prompt for any non-work-profile apps
since the dialog strings today are work-profile specific.

Bug: 326572042
Test: Tested by flashing changes locally with a private space present on
the device.
Flag: android.os.allow_private_profile

Change-Id: I518123b2ee9039791e4da33ee87ed80fd482d272