Reload pointer icons from the UiThread
We're still seeing crashes resulting from frames of an animation
icon having different sizes while loading. The most likely explanation
for this happening is that the display density changes midway through
loading the animation icon, at which point size of the scaled animation
frame will change.

We are already loading the icon from the Display thread. Instead, try
loading the icons from the UiThread, which should hopefully not
receive density updates midway through processing a task.

This is a speculative fix.

Bug: 321324470
Bug: 325133368
Test: adb shell wm density X
Change-Id: Ic2b903d5fdec534aeb32530b124d3b699c40ac33