DSL to define the overscroll behavior of a scene (1/2)
If the overscroll behavior is not specified in the transitions, the
previous transition will be used, ensuring backward compatibility.

The definition of overscroll assumes a gesture from top to bottom or
from left to right. If a gesture is made in the opposite direction
(indicated by the distance sign), a negative progress will be used.

Test: atest TransitionDslTest
Test: atest SceneTransitionLayoutStateTest
Test: atest ElementTest
Bug: 291053278
Flag: NA
Change-Id: Ib946a4def8d8c7dd5ce0ea7ca4ba28d8b9397f16