ChannelEditorDialog: use @dimen/large_dialog_width
ChannelEditorDialog (which uses notif_half_shelf as the root layout) was
displaying with layout_width wrap_content, which meant the dialog was
snugly wrapped around the app icon, name, and global notification
toggle.

We already have a convenient resource, @dimen/large_dialog_width, for
dialogs like this, so use it!

Bug: 309133675
Flag: NA
Test: manual
Change-Id: Ic3871b721b9ecf37abc6d92b0378003806f20ff2