Add Ctrl+Shift shortcut to change the IME
Currently the Ctrl+Space shortcut will be used to change the IME,
but Ctrl+Shift is commonly used in other platforms. This kind of
shortcut would have interfered with game controls, but after the
Control+Space shortcut was moved to `InputMethodService`, that's
no longer a concern, which allows this shortcut to be added.

Bug: 327198899
Test: Flashed on device
Flag: ACONFIG android.view.inputmethod.ctrl_shift_shortcut DEVELOPMENT
Change-Id: I05e43334f55d5b40bc4309fd382f7336ec2bec8d