[Sb] Properly inflate the estimate view
The previous CL created the estimate view, but never added it to the
parent view, and thus was never shown.

Also, this CL slightly reworks the logic for showing percent, since the
slow case is still done on the main thread, but we can skip it in some
circumstances.

Test: manual
Test: BatteryMeterViewTest
Bug: 314812750
Flag: ACONFIG com.android.settingslib.flags.new_status_bar_icons DEVELOPMENT
Change-Id: I5ad19ac9d95a81a0b149dbf4b0b420d47092f979