[Autofill Framework] Remove confusing logs of AFAA
In local, logcat would print the activity name before the log so developer knows which activity is denied. But in bug report, the line doesn't have an activity name so sometimes developer looking would think autofill is denying the activity while in fact, another activity is denied. Most of the times the denied activity is the launcher activity.

Bug: b/328126985
Test: atest CtsAutoFillServiceTestCases
Change-Id: If810e21a598a26b7a10548b915ef7f581e151bd7