[Flexiglas] Split shade - UI layer
UI layer changes to support the split shade.

Split shade is the single panel shade that's shown for large screens and
unfolded devices where quick settings is one the left and notifications
is on the right.

Includes changes to the view-models for the various
scenes. Because split shade is a single panel, we only use the Shade
scene and do not allow navigation to the QS scene.

Adds a side-by-side layout to ShadeScene and displays a rather broken QS
on one side and a rather broken notifications stack on the other.
Followup CLs will fix both and add much-needed transition animations.

What's broken and missing:
- Notifications aren't showing
- The second page of QS tiles peeks into the first
- Scrolling inside the QS editing mode doesn't work
- QS tiles seems unresonsive and don't update themselves properly
- Didn't test media
- No transition animation work yet
- When leaving the scene while in QS editing mode and returning to the
  scene, we're incorrectly still in editing mode
- Other things

Bug: 328473018
Flag: ACONFIG com.android.systemui.scene_container DEVELOPMENT
Test: unit and integration test for view-models
Test: manually tested on a foldable to see folded and unfolded more
quickly. The current state is https://screenshot.googleplex.com/BzzBhdwwWX5CdAt
Change-Id: I4dbace8238043677e9e2ff9b47a1f9f87f5e3ddb

Change-Id: I56575230d36c92105ef6ee4c506c5ad126711ef8