Add a unit test for switching bubbles
Adds a unit test to BubbleStackViewTest to verify that bubbles are ordered correctly after selecting a different bubble.

This change also moves BubbleStackViewTest away from @UiThreadTest because otherwise we can't wait until animations complete without also blocking the test thread.

Additionally, update PhysicsAnimatorTestUtils to add a callback to be notified when an animator is created rather than overriding the factory.
When multiple tests call prepareForTest, they each override the factory, adding another layer to it and we end up creating a deeper and deeper stack.

Bug: 323310200
Test: atest WMShellMultivalentTestsOnDevice:BubbleStackViewTest
Change-Id: I5371c2c68b82f92e7f2d824f40e75ec668505216