Fix QS Edit scrolling in Split shade.
When in customizer Swiping up should not transition to a different
scene. Maybe it should if swiping on the notifications side or from the
bottom, but this is not supported (along with the nested scrolling).

Also, fix the scrolling on the QS side in split shade. The footer should
be fixed and not scroll with the rest of the content.

Test: manual
Test: atest ShadeSceneViewModelTest
Fixes: 329662922
Bug: 330200163
Flag: ACONFIG com.android.systemui.scene_container DEVELOPMENT
Change-Id: Ic69e15fbd402b75dccc75fa7812bcd7dc23412db