Fix qsFullScreen when QS only partially expanded
The qsFullScreen flow & value were true as soon as the QS panel started
to expand, although the JavaDoc said it should only be true when
qsExpansion is 1 (which is true in the flexiglass implementation).

This fixes that by only setting it to true when QS are fully expanded.
This means it's being updated more often, so I also added a check in
NSSL to ignore duplicate updates.

Bug: 330143161
Flag: ACONFIG com.android.systemui.notifications_footer_view_refactor STAGING
Test: QuickSettingsControllerImplTest
Test: NotificationListViewModelTest
Change-Id: I1a2b4244df5199243d954dd1cd4e950e01d7d7a6