Remove spurious calls to the framework when setting selected device
Previously, every new MediaDevice created in InfoMediaManager would
needlessly obtain the current active routing session.

Also, this change removes a potential race condition between the state
of the routing session used for creating the routes list and the state
of the routing session used for finding out the selected device.

Bug: 330300516
Test: Manual.
Test: atest com.android.settingslib.media
Change-Id: I4018dbee599d1ebab6e4ac83ce92e55929494565