Biometric Credman Layered UI Bugfix
When the biometric diaglog was canceled, the UI underneath would remain
awake. This was the first attempted fix, where we listened to a
cancellation error code, and then upon receiving it, given it ensures
the end of a flow, we utilize a viewModel lambda that ensures the
credman UI also ends.

This fix works in practice.

Bug: 331791049
Test: Built and Tested Visually
Change-Id: I076d19e2361ca1017aa57b32150d361eac0ccb97