[flexiglass] Show bouncer when needed.
Hooks up the legacy StatusBarKeyguardViewManager to switch to the bouncer
scene if Flexiglass is enabled when it's told to show the bouncer.

Fix: 330670513
Test: manually verified that long-pressing on a QS tile brings up the
bouncer and that entering the right credentials then opens up that QS
tile's dialog
Test: manually verified that touching the QS edit mode button brings up
the bouncer. Unfortunately, it crashes with
https://paste.googleplex.com/6608650226892800 after the right
credentials are entered; a separate bug will be opened for this
Flag: ACONFIG com.android.systemui.scene_container DEVELOPMENT

Change-Id: I05d56d88ae1da0566e955a6846894956f8f432a2