[4/n] Disable cts tests with freeform activities when desktop windowing enabled
Desktop Windowing is currently only available on selected devices when the desktop windowing flag is enabled. Desktop windowing currently changes the behaviour of freeform activities and thus when this feature is available the legacy behaviour of freeform activities will not be respected and any tests that depend on this behaviour will fail. So we ignore these tests for tablets if desktop windowing is enabled while we work to have the legacy freeform behaviour and desktop mode coexist. Legacy behaviour of launching activities with multi display also seems to be impact. These tests will also be ignored under that same condition.

Bug: 330152508
Bug: 331170652
Bug: 332937436
Test: atest WmTests:ActivityStarterTests
Change-Id: Iab4a238126c339f0ba07883fdd071c6583f7383a