Add operation to pin TaskFragment
Before, activity can't be launched on isolated navigated taskfragment
unless it's launched from the isolated navigated one.

In this CL, we changes the behavior of isolated navigaton a bit:
If a TaskFragment is isolated-navigated, the TaskFragment and
TaskFragments below the TaskFragment can't be the launch target.

We also add a new operation to pin a TaskFragment.
If a TaskFragment is pinned, it should be on the top of the host
task unless a newly launched activity can't be embedded, or it's
launched from the pinned TaskFragment.

Test: atest WMJetpackUnitTests
Bug: 243518738
Flag: com.android.window.flags.activity_embedding_overlay_presentation_flag

Change-Id: I307b84c8f9e8c2114a7487d581fff97e59eba6b5

Former-commit-id: dbeba7768c7ec6abcd1536b506b5c465654df31f