Introduce Forward Override Options for the stable config
This change introduced a forward overriding flag when the stable
configuration behavior is not by default applied to an app. The device
owners can override the flag for apps that are ready for the stable
configuration behavior to provide a better user experience.

This will apply if the system overall stable configuration is not
enabled, or it is enabled but the app is not targeting SDK level 35+
yet.

Bug: 327313645
Test: Change the override the flag value and check the config
Test: ActivityRecordTests
Change-Id: I9f4bbc23b07f63cb40070c6240b0b37660963860