[dVRR] Allow the intermittent update to set frame rate category to NORMAL immediately
Some Views could vote in a very low frequency (e.g., YouTube Short).
With such use case, it would take a while for the ViewRootImpl to set the frame rate category to Normal since we need to wait for mFrameRateCategoryHighCount and mFrameRateCategoryHighHintCount to be 0
before we can set frame rate category to NORMAL.

This CL allows the intermittent update to set frame rate category to NORMAL immediately. Since we have a relatively restricted criteria to detect intermittent update, it's less like to expect the frame frate category fluctuation
after this change.

Bug: 337933990
Test: atest ViewRootImplTest / atest ViewFrameRateTest
Change-Id: I2a61977fa892a12900f60f27a062614e9cca407f