Do not require keyguard message's min time if being update to empty
For example, don't require that the "Unlocked by face"
message shows the minimum show time if the biometric
message is being update to show nothing. This can happen
if the device starts dozing and the user quickly wakes
the device again. We don't want to reshow the unlocked
message when the user rewakes up the device within the
min show time since the unlocked info is no longer valid.

Test: atest KeyguardIndicationRotateTextViewControllerTest
Fixes: 228414801
Flag: None
Change-Id: Icc593319b4b717f83843d4eee3c8245436b835b6