[flexiglass] Change instantCollapseShade to use snapToScene
This is something that would actually work and not create the flicker
issue described in the attached bug.

Fix: 330672236
Test: manually verified that opening settings from the footer of the
expanded QS scene, when the device is unlocked, no longer flickers the
QS scene at the end of the Settings activity launch animation.
Flag: ACONFIG com.android.systemui.scene_container DEVELOPMENT

Change-Id: Ic339d5a70f866e0110dc2518f18441d09b2d7f5e

Former-commit-id: b334d21fce0f61f80040b8e83aaa7209793f4b2a