Let IMMI#transferTouchFocusToImeWindow() take userId
RemoteInlineSuggestionViewConnector is already a per-user instance,
but when it gets call back

  InputMethodManagerInternal#transferTouchFocusToImeWindow

the user ID has not been propagated.

As a preparation to make the following instance fields per-user, this
CL makes it take the user ID as an explicit parameter.

The user ID is not yet used in this CL thus there must be no
observable behavior change.

Bug: 305829876
Bug: 339358344
Test: presubmit
Change-Id: Idf0cdd3cd56701a4343c6f4faa954605f03e4f31

Former-commit-id: 1ec9394cb129424cae4d5c0ecbb4d3a2deb24b4e