Remove dependency on gate_fgs_timeout_anr_behavior flag.
Now that the enable_fgs_timeout_crash_behavior flag is completely rolled
out, remove any dependencies on the stop-gap flag we had in place. If we
need to rollback the crashing behavior in the future, we can simply
disable the crash flag since that behavior is decoupled from the timeout
callback flag.

Bug: 330399444
Test: atest CtsFgsTimeoutTestCases
Change-Id: Ifd64d11493d60ec4841e13163caa9e08e03e0a89

Former-commit-id: 149611da3cf2b9e57413dd640b0fde2f34549d1b