Use monochrome app icon for notifs
In the shelf and status bar, we want to experiment with showing the
monochrome app icon if it exists.

I'm using the same method as in GroupHelper to fetch this icon.

This should be done on the background thread since it requires a binder
call, but I'll make that change in a separate CL.

There's a known issue similar to b/329091967 where the calendar icon
doesn't get cropped at the bottom.

Bug: 335211019
Test: IconManagerTest + tested manually with a couple of apps (gmail,
play store, calendar)
Flag: android.app.notifications_use_monochrome_app_icon DEVELOPMENT
Change-Id: Ibc12ebb9a6a5a3d3619ade7c009c7aae24913fd9

Change-Id: Id8ebf0668769e013c496b22eddb047f01b880a9d

Former-commit-id: 018dbd83a901e85e4e513c2ff182a5ec975e7629