Add lock screen shortcut for opening glanceable hub
This is added temporarily behind a flag for UX testing purposes and is
not intended to be a proper, full implementation. The flag will not be
advanced, only modified locally via adb command.

Shortcut is on the right by default, it can be moved to the left,
replacing the home controls shortcut, by changing the system setting
open_hub_chip_replace_home_controls to 1.

Bug: 339667383
Test: N/A
Flag: com.android.systemui.glanceable_hub_shortcut_button
Change-Id: I51df40df8274a6f77418543783e1b3014bec473f