Treat manual zen rule more like automatic rules
And store policy and deviceeffects info in the manual
rule rather than directly on the config.

This means the manual rule will always exist, and its
condition will be used to determine if it's active or
not.

This cl continues to store the original config policy
values even when the flag is on so it's possible to rollback
the flag without losing user settings.

Test: ZenModeHelperTest (esp the logging tests)
Test: ZenModeDiffTest
Test: ZenModeConfigTest
Test: NotificationManagerZenTest
Bug: 333530553
Flag: android.app.modes_ui
Change-Id: I961d28f01affdcfae9bbf69b16699137188e50e3

Former-commit-id: f9996319a9249203facaca5a3480590d072720f6