[Divider] Fix divider flicker for cold launch cross-app embedding
The flicker happens because the divider is drawn before the open
transition is completed. When the TaskFragment and the decor surface are
created in the same WCT, we get the callback that the decor surface is
created, and all the other conditions are met to draw the divider. But
we don't have a good signal about whether the transition is completed
and when the secondary container content is drawn. This is especially
significant for cross-app cold launch because it is slow. The flicker
also has a small chance to happen for same-app launch.

This change moves the decor surface visibility operations into post
transition callback to ensure that the decor surface presentation
changes deterministically happen after the transition is completed.

Bug: 343973599
Test: atest TaskFragmentOrganizerControllerTest TaskTests and manual
Change-Id: Id8525d44817f2f530cbeafe458d02ed39a81af0d

Former-commit-id: 71c0761d2315e79725ec73577b1d6360eae1b2a7