[VRR] Disable frame rate boosting with content velocity
As requested, we want to avoid boosting when users start scrolling such
as what we did for touch boost.
In this CL, we also make the value of mIsFrameRateBoosting to false
when content velocity is detected.

Also, add a new variable FRAME_RATE_BOOST_TIME for the frame rate boost
use case other than touh boost

Fixes: 343503604
Test: atest ViewFrameRateTest, atest ViewRootImplTest
Change-Id: I8e53cb365740563030a61bf6c1f459c231a2d84f