MotionEvent: Ensure addBatch works in the current coordinate space
The PointerCoords added via addBatch() are expected to be in the
"current" coordinate space of the event. If the event was transformed
before the batch was added, it should not affect the coordinate space of
the current batch.

To ensure coordinate consistency, we need to transfom the newly added
PointerCoords by the event's inverse transform so that we get the same
values out of the event.

Bug: 342349872
Test: atest MotionEventTest
Change-Id: I1a25fd6bf0f600b77b3c35cf2ef6d567f996b5dc