Move dream status bar to ambient package
This change moves the dream status bar view and controller to the
ambient package so that it can be re-used outside of the dream - on the
hub.

This change also includes a small change to move from using the
WifiRepository in the controller to the interactor, to avoid violating
the rule that ui elements should not depend on the data layer. This
resulted in some kosmos changes being needed also.

Bug: 319728690
Test: atest DreamOverlayAnimationsControllerTest
Test: atest DreamOverlayContainerViewControllerTest
Test: atest AmbientStatusBarViewControllerTest
Flag: EXEMPT refactor
Change-Id: Id020b01810c98a4c7815cd7d387becba51c27bf5

Former-commit-id: d84584531a9addb6b7b02a578cee890713daff61