Add onSignificantPlaceCheck() API
Add API that will be invoked whenever anyone checks if the device is in
a significant place. This will allow us some latitude in implementation
of SignificantPlaceProvider (for example by avoiding geofences if
necessary) in the event of significant power burn without needing to
commit further changes in the framework. This gives us the flexiblity to
fix potential future issues with just implementaion changes. Ideally
this will never need to be used.

Flag: significant_places
Bug: 337870680
Test: manual
Change-Id: If25b328dd8140dedb972489c5363e4d0c7102310

Former-commit-id: f77ce1ca16f50414d42089c506b755317594feb0