Ensure all public views receive the group when.
Previously this worked only for public notifications that had a notification header, which was only the default public view, but it wouldn't work for apps which provided a publicVersion, because those would use the headerless collapsed style.  By finding the view time view on the entire wrapped view, we can ensure it will work as expected.

Fixes: 336187519
Flag: EXEMPT trivial bugfix
Test: atest SystemUITests
Change-Id: I48608e01343c79163d3495329bd0ecee7d3d2ef9