[SB][Screen Chips] Refactor OngoingCallRepo to use model instead of boolean.
In order to show the ongoing call chip with all the other new chips, we
need some additional information from the OngoingCallRepository besides
just "whether there's a call or not". This updates the repo to return a
model representing the call state instead of just a boolean.

Also, it seems OngoingCallController + Test needed a full reformatting
:D

Bug: 332662551
Test: atest OngoingCallControllerTest OngoingCallRepositoryTest
StatusBarModeRepositoryImplTest
Test: With com.android.systemui.status_bar_screen_sharing_chips flag
off, start an ongoing call -> verify old chip shows up. Open a
fullscreen app and verify the status bar still appears with a
semi-transparent background (that's the only thing the
OngoingCallRepository is actually used for at the moment)
Flag: NONE Refactor that would be more confusing if it was flag-guarded

Change-Id: I35bd5fd730797b42214e6aa2cc034dc5da7d96e5

Former-commit-id: c4c7c0f67cd98b39b2f511f341b4f44883286780