[Divider] Update veil visibility depending on if container bounds are empty.
If this is not done, the veil surface could be sized incorrectly, and we want to ensure it is not shown. e.g. If a zero window crop is applied, the veil may expand to beyond the full screen bounds (though the view on top may or may not cover it): https://screenshot.googleplex.com/5xiAsixgZSGiBJv.png

Bug: 343509560
Test: atest DividerPresenterTest
Change-Id: I41436d0388b3692293121f7822df5512f988bc56

Former-commit-id: 830f2d20e4452f412299e94f8752beb239bc3cb9