Add system property to disable BIC
System property is added for different form-factors to disable
Background Install Control service.

The feature flag enableBicFeatureGuard has also been added to roll out
this change.

Looking from source, the only two system component that is related to
bic are BackgroundInstalledAppsPageProvider in Settings (which is not
used on Wear) and BinaryTransparencyService, which has already guarded
for the case where getSystemService for bic can be null. Thus this
change should be safe to land.

Bug: 340928990
Flag: com.android.server.flags.optional_background_install_control
Test: Local build + confirm bic not strting + presubmit
Change-Id: I3814a6171e2fdfcc575420a2722c719462bf3be0

Former-commit-id: 6aca162fa7f7d066da9e698674487e74332847f4