Expose startedPosition and pointersDown to NestedScrollHandler
Until now, we have handled scrolls as gestures without a specific
starting position (using null) and have only considered one finger on
the screen.

Moving forward, we aim to accurately expose this information and we can
achieve this by leveraging the PointerInput API.
The NestedScrollToSceneNode now employs a PointerInputHandler to
determine the number of pointers and obtain the starting position during
 the nested scroll.
The PointerInputHandler awaits the creation of the
AwaitPointerEventScope to register the nestedScrollHandler.
This approach ensures that touch events are always received by the
PointerInputHandler first, in the Initial step, and later these events
are consumed by our descendant scrollable in the Main step.

Test: atest ElementTest
Bug: 330200163
Flag: com.android.systemui.scene_container
Change-Id: Ib912aec5af40145af7a0d580aa81b3098d0ad092