Do not show unavailable tiles in edit mode.
Filter out tiles that are unavailable (cannot be added). For new tiles
we use QSTileAvailabilityInteractor (part of the QSTileDataInteractor),
but if the flag is off or the tile has not been migrated yet, we default
to constructing the tile, checking its availability and then destroying
it (similar to TileQueryHelper).

Test: atest com.android.systemui.qs.panels
Fixes: 339262140
Flag: com.android.systemui.qs_new_tiles
Flag: com.android.systemui.qs_ui_refactor

Change-Id: I1e91a6c1ab0185d201bfe853af223363280f8ec4

Former-commit-id: f5598a57a6f7d7b9a8d67e33d255b1c8e111a556