Temporarily remove @Nullable from findViewById(int) methods
Previously, Metalava had a nasty hack that would ignore the `@Nullable`
annotation present in the source code definition of any
`findViewById(int)` methods. This change explicitly removes the
annotations from the source code which will allow the Metalava hack to
be removed.

Longer term, the calls to `findViewById(int)` that would be broken
because they treat the returned value as being `@NonNull` will be
replaced with `requireViewById(int)`.

Bug: 347204084
Test: m droid
Change-Id: Ie67b40dd3321f403bcecffddd11453038b2329a9

Former-commit-id: de08dbd7088163e095238598f9fc85c6b93702c1