Cleanup IME subtype switching controller and tests
This adds annotations for the nullability of parameters, values and
return values of methods in the IME subtype switching controller, as
well as simplifying the existing logic with an equivalent one.
Additionally fixes the testImeSubtypeListItem to check the
isSystemLanguage of all items, as well as add the missing item_en_uk.
Lastly, replaces the non-inclusive language from the test class.

Test: atest InputMethodSubtypeSwitchingControllerTest
Bug: 347693610
Change-Id: I6292fd34f016d7a1635994a364d91eb071c5d5b5

Former-commit-id: a1e9a88281e2502198ae2035f1249bb351c51410