Displaying the base activity's icon and app label in app header
We are currently using the top activity to get the icon and app label to
display on the task header. When another activity corresponding to
another application opens in the task, when we create the header, it
will display that information rather than the information of the base
application. This change gets and derives the info from the base
application instead.

Test: Open better bug and transition to freeform. header should have
name and icon of the better bug app.
Bug: 341992887
Flag: EXEMPT bugfix

Change-Id: I171c1113ceebdcfe594554eeef5679c651047fe8

Former-commit-id: c40b465f3058218914b2f292ce433ecbbb85abce