Temporarily remove @Nullable from getSystemService(...) methods
Previously, Metalava had a nasty hack that would ignore the `@Nullable`
annotation present in the source code definition of any
`getSystemService(...)` methods. This change explicitly removes the
annotations from the source code which will allow the Metalava hack to
be removed.

Longer term, the calls to `getSystemService(...)` that would be broken
because they treat the returned value as being `@NonNull` will be
replaced by a new `requireSystemService(...)` method.

Bug: 347269120
Test: m checkapi
Change-Id: I9709acfb57c6954401ea617329c821cb4efa0889

Former-commit-id: 30540639b0e13a384f52580f81a4ef520b6d5084