Drop the first keyguardAuthenticatedBiometricsHandled evaluation
Since the event is derived from a StateFlow, the starting state
is always null. The transition from alternate bouncer to gone
should only be triggerd on subsequent updates to the value.

Flag: EXEMPT bugfix
Test: atest FromAlternateBouncerTransitionInteractorTest
Test: enroll UDFPS, set a timer for 10s, go to lockscreen
and manually trigger the alternate bouncer, wait for
the timer to go off (and show the clock activity), navigate
back to the lockscreen and observe UDFPS icon is visible
(and check keyguard transition states were updated correctly)
Fixes: 345373593

Change-Id: Idb1c077e993345bf04d38ebee7615cf5429e25f8

Former-commit-id: 0c969e275208394f77230041097861f5f3c84468