Remove use of --check-compatibility:base
The `--check-compatibility:base` was added before multiple instances of
`--check-compatibility:api:released` were supported. It was also
inaccurate in a couple of ways:
1. It used the same old base file as the base for both the old and the
   new. So, it was comparing "old base + old" against "old base + new".
2. It did not include the system delta when checking `module-lib`.

This change removes the `--check-compatibility:base` and replaces it
with:
1. Additional `--check-compatibility:api:released` to specify the full
   previously released API surface.
2. Additional `frameworks-base-api-*-current.txt` files to specify the
   full combined API surface.

It did not use the `...combined...` file groups for the former because
there was not a clean way to prefix each file in the group with its own
option. It is possible to use scripts to generate that but it would be
more complicated than just listing them out individually.

Bug: 320245343
Test: m checkapi
Change-Id: I8c2a288ec9deeeab44f56758931a5f97b932c929

Former-commit-id: 4b65c4ced0781d281cdd7ad72d3aeeb8019086b0