Fix InputMethodSubtypeSwitchingController tests
When checking the switching-aware loop order, we only use 3 items, with
an user action on the second item, and then on the third item. Due to
the circular nature of the order, an user action on the last item
results in the same order. This adds an additional subtypeAwareIme to
ensure the order is modified and checked accurately. Additionally this
inlines the systemLocale parameter of createTestItem as it was always
the same.

Flag: TEST_ONLY
Test: atest InputMethodSubtypeSwitchingControllerTest
Bug: 347693610
Change-Id: I5f8c9519713b1731b46fb809df80db8b3deb289a