Clear active sessions when stopping touch monitoring.
It is currently possible to stop TouchMonitor even if there are active
touch sessions. This is needed when there is a lifecycle change and
touch handling must be disabled. While the sessions are currently
notified and cleaned up, TouchMonitor fails to clear the active session
references. Therefore, future touch events are sent to essentially
empty shells of past sessions.

This changelist addresses this issue by clearing the list of active
sessions after each has been notified.

Fixes: 348656793
Test: atest TouchMonitorTest#testSessionResetOnLifecycle
Flag: EXEMPT bugfix
Change-Id: If8abd912abdd177932c94fdd84f5e27ac0a53bf3

Former-commit-id: a5d0d2c73df47c7941ec6cfcdffcde7ba38f6232