[flexiglass] Wake up from AOD when lockscreen tapped.
During the first ~4 seconds after entering AOD, tapping the lockscreen
should wake up the device and exit AOD. This wasn't happening in
Flexiglass. This CL makes it work.

The PulsingGestureListener is responsible for translating taps and
double taps into the proper wakefulness state changes. This CL hooks it
up to the pre-existing LockscreenLongPress (which is renamed to a more
generic name in the follow-up CL).

Bug: 344879669
Test: manually verified that tapping or double tapping on the AOD
lockscreen during the first 4 seconds properly wakes up the device and
exits AOD.
Flag: com.android.systemui.scene_container

Change-Id: I44352c788e841a475f5b31a460c69cdcda91685f