Allowing clicks on the IDLE state of the QSLongPressEffect
When using Talkback, clicks are delivered upon a double tap on the
screen. These trigger the onTileClick path in QSLongPressEffect when the
effect is in the IDLE state. This is also a valid click and it is now
allowed in this state.

Test: atest SystemUiRoboTests:QSLongPressEffectTest
Flag: com.android.systemui.quick_settings_visual_haptics_longpress
Bug: 348295719
Change-Id: I12b8765fcc154b7d83ccf86e3a4ca33cca264644

Former-commit-id: 0c973ce7d48f6c4b1fae1ca840caa62ec1aaeff7