Flexiglass: hydrate dozeAmount from transitionValue(AOD)
To animate between the lockscreen and AOD, we need to know whether we're
dozing (AOD or off), and how far we are into that state (dozeAmount).

With Flexiglass off, NotificationPanelViewController will call
StatusBarStateController when it needs to animate into or out of doze,
and SBSC will drive the animation itself, updating dozeAmount as it
progresses.

With Flexiglass on, NotificationPanelViewController is out of the
picture, and the animation itself is driven by the keyguard transition
framework, but (until this CL) we didn't have anything updating
dozeAmount.

This CL makes StatusBarStateController subscribe to updates from
KeyguardTransitionInteractor and update dozeAmount accordingly.

Bug: 347731926
Test: manual: turn on AOD, go to lockscreen, turn screen off/on
Flag: com.android.systemui.scene_container
Change-Id: I80d439904abec97fd2f1b8b29f351706dcf297ad

Former-commit-id: 7ce64b3ecdb5581ebaf5a5991d71d1f4053390bf