Fix lastDispatchedState sources change
Before, in InsetsController#updateState, it directly used the same
source reference to update mSourceConsumers and mState, which will then
get the local override. This can cause issue when the activeControl is
revoked, because it will only trigger notifyInsetsChanged if the local
and last dispatched are different.

Bug: 339380439
Test: WindowInsetsControllerTests#testImeForceShowingNavigationBar
Flag: com.android.window.flags.insets_control_seq
Change-Id: Ibadeca3b6bbfa46f1e2fd0adfaf3e63b6abe8b3f

Former-commit-id: b5600f5f9741f9abf11897e237390843c6343024