[SB][Flexi] Make KeyguardStatusBarVCTest pass regardless of scene state.
Previously, if you had the scene flags enabled on your device, then a
lot of the KeyguardStatusBarViewControllerTest tests would fail because
many KSBVC methods are disabled when scenes are enabled, but the test
methods didn't check the flag.

This CL adds @DisableSceneContainer to all the tests that are testing
methods that don't work when scenes are enabled.

Bug: NONE
Flag: TEST_ONLY
Test: atest KeyguardStatusBarViewControllerTest with all scene flags off
Test: atest KeyguardStatusBarViewControllerTest with all scene flags on
Change-Id: I3bb00084e3769466f5a0df60e527088340f795b3

Former-commit-id: f0bd1e59f28bc3c13dbeb9a259ef9b06065e814b