Add user serial number to communal db schema
This change adds a new column to the communal database storing the
serial number of the user associated with each widget. It also
implements a migration that is run when a device is updated from the
previous version of database, populating the default value as -1.

Test: atest CommunalDatabaseMigrationsTest
Test: atest CommunalBackupHelperTest
Test: atest CommunalBackupUtilsTest
Test: atest CommunalWidgetDaoTest
Test: atest CommunalWidgetRepositoryImplTest
Bug: 330945203
Flag: com.android.systemui.communal_hub
Change-Id: I93c231ab9a25f580730628815e588e986ffd0102

Former-commit-id: 4986ebfb69a3b0254a655249f68086906da9bb83