Remove hasCommittedReparentToAnimationLeash
For BackNavigationController:
It was added in commit 642b0d8 to avoid conflicting with manual
created leash, but commit 312da01 has changed it to regular
WC#startAnimation which will manage the leash by SurfaceAnimator.

Also the mPendingAnimationBuilder will run on onTransitionFinish,
but hasCommittedReparentToAnimationLeash doesn't mean that there
is a transition. That becomes an unpaired condition, so it should
not check that to set pending.

For AbsAppSnapshotController:
It was added for very old split screen when dismissing one side
(commit 7f1fa99). Since shell transition, there is no longer a
leash in core, so the condition for transition never satisfied.
And even the snapshot can still be taken successfully for the
original dismiss-split case.

Bug: 163976519
Flag: EXEMPT unused logic
Test: CtsWindowManagerDeviceActivity

Change-Id: Ic61e5736d7ad85064ce5f43edcd4c272d406c612

Former-commit-id: 2722e5ce99daa4eac468b438669cd57514c07ef3