Re-land^3: Let WindowProcessController apply app specific config
Since the feature of insetsDecoupledConfiguration, targetSdk can affect
the app sizes of the configuration. An app might get the configuration
from WindowProcessController before the size is overridden by the
ActivityRecord.

This CL applies the size override to WindowProcessController as well so
that apps won't get the unexpected apps size from the configuration.

Bug: 347136833
Bug: 151861875
Flag: com.android.window.flags.insets_decoupled_configuration
Test: atest WindowPolicyTests QSMediaPlayerContent
      WindowProcessControllerTests FrameworksMockingServicesTests
Change-Id: I033ef2c3dda39ac2d69c553c3a83d1b18e86fa08

Former-commit-id: 0b58b352fd9947479b5fc72634392f2d33c0c117