chore(magnification panel): make size and diagonal title multiple line
Previously for b/280162065 and b/328519260 we change the textviews to single line marquee mode as a quick workaround. Now we find that for panel's LayoutParams width, if we use a runtime calculated width instead of WRAP_CONTENT, we can solve the multiple line height calculation issue. Therefore we'll change the textviews back to multiple line mode so that users can directly see all the context.

Bug: 286479523
Flag: EXEMPT bugfix
Test: manually
      atest WindowMagnificationSettingsTest
Change-Id: I45a2ed75de73b0916a580c2047678a6d95b4c026

Former-commit-id: f9ddb9b91fdfe6947c75de33df99e315c48123df