Listen for ongoing updates when communal enabled
This change makes it so that updates for ongoing content (media and
smartspace timers) are listened to as long as communal enabled.
Previously, media updates are always listened to, even when not
applicable. Smartspace updates were only listened to when communal was
showing, leading to delays and unexpected behavior when connections were
closed and reopened.

Test: atest CommunalOngoingContentStartable
Test: atest CommunalSmartspaceRepositoryImplTest
Test: atest CommunalMediaRepositoryImplTest
Bug: 322869039
Flag: com.android.systemui.communal_hub
Change-Id: I10c02d0a1e54d2dddef4f7979cdbee6fd0ba0efe

Former-commit-id: 5e7a6fa6395209cc3d6438e0c338918bd9d1ce16