Remove requirement to get registered group on tracing start and stop callbacks.
We get these callbacks on all processes with the same config. So in some processes the group will simply never be registered leading to an exception.

We can't simply ignore groups that are not registered in the processes by the time we start tracing because the registration might happen after the trace instance starts. In which case we still want to track the config for when this group ends up being registered.

The solution here is to use String identifiers instead of the group object as keys where required.

Flag: android.tracing.perfetto_protolog_tracing
Test: Run Perfetto AOT config to collect a trace
Change-Id: I3c85fed402a1c9eefc55f82d0cef0705190d4afe

Former-commit-id: 8a5554d79be2976f43a8ccf88d93157ac749d038