Do not skip IME in forAllWindows if its target is invisible
This removed the logic added in [1]. Because forAllWindows can be
used to perform any operation. It is unsafe to skip generally.
Such as when performing applyKeyguardPolicy for all windows, it may
miss the chance to hide IME.

Also the original problem in [1] usually won't happen since IME
surface can be attached to app's surface, then the IME can be hidden
with the app naturally.

[1]: I57357ba85501397fa5926ab4dee116c42df24506

Bug: 339375944
Flag: com.android.window.flags.do_not_skip_ime_by_target_visibility
Test: WindowContainerTraversalTests#testTraverseImeRegardlessOfImeTarget
Change-Id: I73b952f384ec1f6104e548dad62a0483c6f5915d

Former-commit-id: 0ff4ed6f099b1f04f3aae9ea563a3a0915a650f5