[flexiglass] Fixes issue where LS status bar appears too late
The status bar view would get added too late, causing the layout below
it to be pushed down vertically in a janky manner. This was happening
because of an unnecessary use of MovableElement when no other scene was
actually using the same view.

This CL fixes that by removing the MovableElement.

Fix: 352052842
Test: manually verified that the jump no longer occurs
Test: manually verified no harm done when unlocking to go to Gone and
when pulling down the shade and QS on top of the lockscreen or the Gone
scene
Flag: com.android.systemui.scene_container

Change-Id: I997cf59d805bbefa7dbdfcbfb09f44b1b982a44a

Former-commit-id: ae09793299a584a4cb4f47ee538c7f201802102c