[SB][Screen Chips] Re-add animation from chip -> stop dialog, and CUJs.
It turns out there's a workaround that lets SysUI do a nice animate-in
from view -> dialog, and then do a standard animate-out so that the
dialog doesn't try to animate out into a view that's about to disappear.
This CL re-adds that nice animate-in and includes the workound. It also
adds CUJ tracking.

Fixes: 353249803
Bug: 332662551
Flag: com.android.systemui.status_bar_screen_sharing_chips

Test: start a screen share, then tap chip -> verify the chip animates
nicely into the dialog. Tap "stop" on the dialog -> verify the dialog
does *not* animate out into the chip
Test: All tests in statusbar.chips package

Change-Id: I84993c7135152cf3f50d22975abd97438fb84824