Run IMMS#onUnlockUser() also for background users
As part of our on-going effort to address the remaining dependency on

  InputMethodManagerService#mCurrentUserId,

this CL updates the behavior of

  InputMethodManagerService#onUnlockUser()

so that it calls

  * IMMS#postInputMethodSettingUpdatedLocked()
  * IMMS#updateInputMethodsFromSettingsLocked()

not only for the current user but also for background users, even for
phones/tablets/TVs/wearables. This was not possible until when we
update the above two methods to take userId [1].

With this CL, we would start initializing IME-related settings
(e.g. secure settings) more eagerly for background users. But other
than that, user observable behavior should remain unchanged.

 [1]: I9d4615e6da979204c21555e040a687163c1460ab
      8f23a7dce53620c4697e590fd27cd55de8e5dea1 [formerly e08e5c4a76ea32db05af858ebcfd9d88dc943f8c]

Bug: 350386877
Test: presubmit
Flag: android.view.inputmethod.concurrent_input_methods
Change-Id: Iecbca2696d4e2fd56a434f2e3a95d14976978e06

Former-commit-id: 1c1954f6207b3cd46f9c6525ebb996dfde563126