Refactor WebViewUpdateService context handling.
Previously, a number of the system interface methods took Context
parameters. However, the real update service implementation only stores
a context in order to pass it to these methods, and the mock
implementation of the system interface doesn't need a context at all. If
a context is needed in a system interface that currently doesn't have
one, this requires a change to the interface and tests.

Instead, just store the context in the implementation of the system
interface when it's created. This means we can remove the context
parameter throughout the rest of the code, simplifying the mock/test,
and future changes to the implementation can rely on having a context
available for all methods.

This is a pure refactor with no behaviour changes.

Bug: 310653407
Test: atest WebViewUpdateServiceTest
Flag: EXEMPT refactor
Change-Id: I53bf25ab2d5f9b2636ae279cd8ac72e473e97ee6

Former-commit-id: 3fe6f8d21673fa6f8118a4ac63746779f0af14cf