Consume but ignore glanceable hub touches when entering and exiting edit mode
When entering edit mode, the communal scene is kept open until the edit
mode activity is ready. Similarly when exiting edit mode, the communal
scene is started before the device is locked. This allows for smooth
transition animations to and from edit mode, but also means that the
presence of the communal scene is no longer correlated to the
GLANCEABLE_HUB keyguard state and other important logic.

User interactions during these transition periods can cause a number of
issues, like missing notifications, a missing shade background, landing
on a blank edit mode activity, etc.

To prevent users from seeing these issues, we consume but do nothing
with touches in the hub and pause TouchMonitor during these times. This
is a relatively short period of time but solves many funky issues.

This change also adds noHistory to the edit mode activity so that it's
finished in case the user exits with a back gesture, home button press,
or gesture nav swipe, so that the editModeOpen state updates properly.

Bug: 350815853
Bug: 352171134
Bug: 349385145
Bug: 349386602
Test: atest GlanceableHubContainerControllerTest
      also manually tested on device
Flag: com.android.systemui.communal_hub
Change-Id: I23475cfb3eac15ec2499cfd60424611a2e9a02a5

Former-commit-id: 7ad672957879c5704e12601c0bb2cb2116781bc6