Remove ClientTransaction object pool usages (2/n).
Eliminate the use of ObjectPool in the creation and management of
ClientTransaction objects. Instead of obtaining transactions from
the pool, they are now directly instantiated. This simplifies
transaction handling and aligns with the deprecation of the
ObjectPool mechanism.

This change also removes the unnecessary recycling of ClientTransaction
objects in the ClientLifecycleManager.  As part of the cleanup,
unnecessary code related to local transaction recycling in
ActivityThread was also removed.

Bug: 311089192
Test: atest FrameworksCoreTests:ActivityThreadTest
Test: atest FrameworksCoreTests:ClientTransactionTests
Test: atest FrameworksCoreTests:ObjectPoolTests
Test: atest FrameworksCoreTests:TransactionExecutorTests
Test: atest FrameworksCoreTests:TransactionParcelTests
Test: atest WmTests:ClientLifecycleManagerTests
Flag: EXEMPT removing com.android.window.flags.disable_object_pool
Change-Id: Ib025874410ef09bc2487007642a321c934d52cef

Former-commit-id: e68b5c19232860f88b515f7d022814ea9eded743