Fix OriginalUnseenKeyguardCoordinator e2e test
in ag/28176151 the seen behavior was slightly changed which caused a e2e
test to fail: b/351913295. The intention was to use standardized
flexi-aware APIs and the change seemed to be ok, but the test shows that
there is a timing issue somewhere downstream. This CL reverts the logic
and replaces it with a custom flagged logic that mimics the same
behavior in Flexiglass. There is no flexi-aware API for this because
usually we check being on a state or leaving the state (isFinished) but
not going to a state or being on a state.

Bug: b/351913295
Flag: NONE revert unflagged behavior
Flag: com.android.systemui.scene_container
Test: Revert to old behavior
Change-Id: I840d18b5979cc28a88af88b604f605bbda1698e3

Former-commit-id: 52b08f428e95f1d710ff3d1e900db1eef2806c2a