Unify IMMS#onSecureSettingsChangedLocked() behaviors among users
This is a follow up CL to my previous CLs [1], which made our secure
settings observer multi-user aware.

What is remaining is an early exit case at the top of

  InputMethodManagerService#onSecureSettingsChangedLocked(),

which basically makes the above callback no-op when called for
background users when IMMS#mConcurrentMultiUserModeEnabled is false.

With this CL the above early exit rule is removed, meaning that
updates on secure settings will be propagated into downstream logic
even if called for background users.

 [1]: I35253d3964f37f3e13888c0017a86449cf1de8ac
      78b2ba071a85bc67628c3ca12f1acdfbb5d3f2ad [formerly e49e099480b1cf53caea747c0a799914e02e1591]
 [2]: I27f94ca795f3035b5132a28fbbec669122d42b87
      3fcc5be8c6f6281204a50bb066bc23e206c417c0 [formerly 6cfdf037fa72e80b125858945532a4ca9fa8d56a]

Bug: 342027196
Bug: 350386877
Flag: android.view.inputmethod.concurrent_input_methods
Test: presubmit
Change-Id: I039816fa3e2acfe3cada3fb237c01560d474aee8

Former-commit-id: 572334b67688c122b5a6c852e5b8e73143efb8aa