Allow any non-touchscreen device click to bypass long-press need for shortcuts
Previously only mouse and stylus were able to avoid having to long-press
to activate a shortcut.

Inversely, we only want the touchscreen to require a long-press, which
this adjusts to.

This allows for trackpads, game pads, trackballs, and many other
accurate devices to activate the shortcuts without having to long-press.

Fixes: 355692248
Test: manual - tested on tablet with trackpad attached
Test: manual - tested on tablet with mouse attached
Flag: com.android.systemui.non_touchscreen_devices_bypass_falsing
Change-Id: I175cbf79834b48d21b20ff42320d8889bd9e26a4

Former-commit-id: 61564444db47d9ebd80484d2baaf2f646a269db3