Revert "Remove UserLifecycleListener deps from SecureSettingsWrapper"
This reverts commit fd2bc931dc27fd98af2eb0c2ee5ddf37aaa4041a [formerly 1a7311fe9a04a840f730144d72f12d74f2005511].

Reason for revert:
While it is highly unlikely, reverting this is the fastest and easiest
way to to demonstrate that this CL is not the real culprit of the SysUI
perf regression discussed in Bug 355334507.

Once it is confirmed that this CL is not the culprit, we will re-land
this as this CL should indeed improve our memory consumption.

Bug: 352354308
Bug: 355334507
Flag: EXEMPT refactor
Test: presubmit
Change-Id: Ib4ce3cb283f5165cff0e66c9a6ddc29910abd911

Former-commit-id: 44a859a61eebc24f995c73eb0af96abfe82cf176