Introduce StaticElementContentPicker (1/2)
This CL is a partial revert of ag/27025577, which allowed
ElementContentPicker.contentDuringTransition() to return null to
indicate that the element is neither transition.fromContent nor
transition.toContent. This was needed for movable elements, which needed
to know *at composition time* in which scenes/contents an element could
be.

Once overlays will be introduced, we will still need to be able to do
this check but we won't always have a transition when an overlay is
shown but not animating. This CL introduces StaticElementScenePicker,
which can provide the list of contents in which an element can be. This
StaticElementScenePicker must now be used for MovableElements.

This CL also converts some MovableElement from keyguard sections into
normal Elements. I expect those elements to not be used or not work
correctly anyways because these elements are using the default
HighestZIndexContentPicker, which does not work well with movable
elements that don't specify the scenes/contents in which they can be.

Bug: 353679003
Test: atest PlatformComposeSceneTransitionLayoutTests
Test: atest MovableElementTest
Test: atest MovableElementContentPickerTest
Flag: com.android.systemui.scene_container
Change-Id: I39a0427de2eaccd133bbf206ed435a801e07e539

Former-commit-id: 8ad770b757cb3ef13dbedb77bb1234f7bb695d30