Simplify DraggableHandlerImpl.onDrag()
In this CL, our aim is to simplify the onDrag() function and minimize
object creation during the drag gesture.

Instead of directly updating the swipeTransition.dragOffset, we
introduce a desiredOffset. We then check if we have reached the toScene
and use this information to calculate the newOffset. This approach also
considers transitions that cannot overscroll.

Following this, we update the current transition and potentially create
a new transition, if required.

Test: Just a refactor
Bug: 336710600
Flag: com.android.systemui.scene_container
Change-Id: I501b47b58e2f7e47b2aeb6cb15c3028cbb422877