Add FLAG_SPLIT_TOUCH to window decoration windows
Adds it to the App Header, App Handle, both menus and the
DragResizeInputListener as a workaround for a crash when splitting
motion event is enabled.

Also removes the input channel from the ResizeVeil as it is not needed,
and since drag-moving two tasks simoultaneously with two pointers is now
possible, this change prevents |moveTaskToFront| from being called on
every |onTouch| call for each of the tasks, which would cause flickering
as both tasks tried to move to front repeatedly during the dual
drag-move gestures.

Bug: 352306102
Test: check input continues to work on all of these windows, and
drag-moving two windows at the same works fine without flickering.
Flag: EXEMPT bugfix

Change-Id: I54630b1d26444f31a52a47c2756ebb90d5d66bbf