Use user context when loading pip menu icons
There is a bug involving trying to load app icons when using secondary
users (or HSUM). Since we load icons using the systemui context which
is always user 0, we might not be able to load the icons if the app
is not installed on user 0.

Fixed by using the current user context when loading menu icons.

Bug: b/341192726
Flag: EXEMPT bugfix
Test: Follow repro steps as listed in BR, verify icons show.
Tested for both secondary user and HSUM.

Change-Id: I9297842ab90b91d4786d37c410f45b372859ee69

Former-commit-id: f6c24aeecc128d4d8a4a40a4e6f17fa3842d40d6