camera2: Fix NPE in SessionConfiguration#hashCode
SessionConfiguration#hashCode currently calls mInputConfig.hashCode()
which may lead to a NullPointerException if mInputConfig is not
set.

This patch replaces the mInputConfig.hashCode() call with
Objects.hashCode(mInputConfig) to ensure that null value of
mInputConfig is properly handled, and indicates that
mOutputConfigurations cannot be null by contract.

Bug: 358211021
Test: Verified that calling SessionConfiguration#hashCode()
      with null mInputConfig no longer throws a
      NullPointerException.
Flag: EXEMPT minor bugfix
Change-Id: I30a7e1cb2372a42d4c3af2e2659bc429111b4dae

Former-commit-id: c21dcc02d9fe7de0e55c3fc587322f63ee86b1c8