Enable switching away from auxiliary/unknown IMEs
Previously any attempt to switch to the next IME while the current IME
is auxiliary would fail, as auxiliary IMEs are not included in the
switching rotation lists, and would early return in these cases.

This enables only switching away from an IME that is not known in the
switching rotation lists.

Flag: android.view.inputmethod.ime_switcher_revamp
Test: atest InputMethodSubtypeSwitchingControllerTest
Bug: 358271932
Change-Id: I079e61e80765970205fed06edd0daaea461a2ff0

Former-commit-id: 922e4fbd706d5032880e768046344b0caacd8bc1