Fix failed testManageSpacePendingIntentNoBalAllowed for secondary_user_on_secondary_display
When a test runs as the driver (current user),
tradefederation will install the testing apk for both the current
user and system user. But when the test runs as the passenger,
tradefederation will only install the apk for the passenge
(see b/350565912), so testManageSpacePendingIntentNoBalAllowed
will fail when running as passenger.

To fix the test, the CL changed BackgroundActivityStartController
to check whether the apk is installed for the testing user instead
of the system user.

Bug: 351081318
Test: atest CtsWindowManagerBackgroundActivityTestCases
Test: atest --user-type secondary_user CtsWindowManagerBackgroundActivityTestCases
Test: atest --user-type secondary_user_on_secondary_display  CtsWindowManagerBackgroundActivityTestCases
Flag: EXEMPT bugfix
Change-Id: Id10740ab29d454ca26113a74dad5c7a710584096

Former-commit-id: 5977ba076f0272da583601f4f9eda8f4c4d83f53