Add hover support for PinPadButton
This CL adds a hover state for the PinPadButton
by using the default behavior in combinedClickable
along with a CircleShape clip.

The existing isPressed animation is kept by using
a null indication when isPressed is true.

Bug: 283802717
Tested: Made sure a hover highlight is displayed using
a mouse with the Pixel Tablet. Also, ensured the existing
UI and animations are unchanged.
Flag: com.android.systemui.scene_container

Change-Id: Ife685063d7b27126e528af81e7ed70211a39e243

Former-commit-id: eba8f9d85b2f6a8f0820d5f45438a6760d3f1db8