Fix stuck transitionValue
There is a race condition happening, with an incorrect
transition starting just before another. Because of this,
a prior transition is getting canceled but the
transitionValue is left in a bad state that is very hard
to get out of. Until the race condition can be resolved,
add some cleanup to get the user out of this state.

Bug: 358533338
Test: atest KeyguardTransitionInteractorTest
Flag: EXEMPT bugfix
Change-Id: I20bed404be1664f0a5d553e740e55e9b50133440

Former-commit-id: 14240ef3e3a20cbcdfcabbeb103948fb73575826