Enlarge mLast4FrameMetricsInfos from 4 to 6 entries
As a result of tuning SurfaceFlinger durations
(debug.sf.late.sf.duration etc.), the ring buffer of the frame metrics
infos is sometimes too small, and already popped the frame's info when
its GPU complete is triggered and doing onSurfaceStatsAvailable(), so
finishFrame() is not called for this frame, which causes incorrect
jank judgement on the next frame.

Increase the ring buffer size from 4 to 6 can avoid the issue.

Bug: 354618175
Test: no skipped frame in onSurfaceStatsAvailable
Flag: EXEMPT bugfix
Change-Id: I2865365213e16cfebe63b949aae332cea32d8927

Former-commit-id: 6ba5125b9d5b88a4e9292f842f401827c0ad55f2