Move ContentState.Transition back into TransitionState.Transition (1/2)
This CL is a revert of ag/28505631: in my first iteration of overlays,
there was a strong split between scene transitions and overlay
transitions, they were contained in separate collections in STLState.
However for swipe gestures I realized that we still need to support the
concept of "isActive", i.e. whether the current transition (scene or
overlay) is the current, last one. This means that we still need some
kind of ordering between scene and overlay transitions, which is why the
second iteration of overlays (that will be introduced in ag/28338922)
keeps the current list of transitions in STLState.currentTransitions for
both scenes and overlays. For this reason, we still need a base class
for transitions, which is TransitionState.Transition.

This now means that TransitionState.Transition is the generic transition
base class (between contents) and not the specialized scene transition
(between scenes). For this reason, this CL also introduces the
TransitionState.Transition.ChangeCurrentScene for the more specific
scene transitions.

Bug: 353679003
Test: atest PlatformComposeSceneTransitionLayoutTests
Flag: com.android.systemui.scene_container
Change-Id: I6c464cdc6211cb7092f7438c2bf24f9a9e5ad603

Former-commit-id: 318638dcd3b61c7be407a75b7a0cf08c1691961c