Refactor startDismissKeyguard into KeyguardDismissTransitionInteractor
There are two issues with that method being in
KeyguardTransitionInteractor:
1. Because it references back to TransitionInteractors it creates
dependency cycles that were resolved by using dagger.Lazy (avoid where
possible)
2. KeyguardTransitionInteractor is supposed to be a source to query or
listen to the state of Keyguard Transition Framework. This violates SRP
by adding unrelated functionality that a) is not relevant to most users
of KTI (usually developers of lockscreen features that don't have deep
understanding of KTF) and b) adds a lot of noise and dependencies to the
class which makes it hard to parse and pick the relevant APIs.

Bug: b/330311871
Flag: EXEMPT refactor
Test: REFACTOR_ONLY
Change-Id: I7b7d8da15d62de6e82e678539b6439e7dff1ea3f