Remove ProxyFileDescriptorCallback in onRelease
Clean proxy FileDescriptor callback from fuse app loop thread when the file is
closed and after onRelease on the callback is called. This guards
against memory leaks that currently occurs as we currently keep the
callback in memory for the life time of the app providing the proxy file
descriptor.

Test: atest android.os.storage.cts.StorageManagerTest
Bug: 211941260
Flag: EXEMPT bug fix
Change-Id: I90ff0159fbc04a86071880bd3b6853bc39e9136d