Do not override bounds when task override exists
This change fixes the issue that sometimes the non-decor insets are
accidentally deducted from override bounds even if the bounds doesn't
overlap with the insets.

A special handle is needed for split screen when the window is pushed
above by the IME. It is necessary to ensure the app receive the same
configuration before and after to make sure the activity won't relaunch
to ensure the IME can be shown correctly.

Bug: 358509380
Bug: 356050755
Test: WindowPolicyTests#testOptOutEdgeToEdgeAppBounds
Test: WindowPolicyTests#testOptOutEdgeToEdgeDisplayMetrics
Flag: EXEMPT bugfix
Change-Id: Ibebb43b29d76e5a9fa81334a6c678aa5f34678b2

Former-commit-id: aaaf1c4749fcb8c4271103794b3135be0b895373