Fix splitscreen with freeform windows
Apps running in freeform window mode may have any arbitrary bounds, and
therefore will not inherit the bounds of their parent root tasks of the
main and side stages, causing splitscreen to be broken.

To fix this, the bounds of these leaf tasks must be cleared when they're
put in splitscreen.

Detailed explanation and demo video of the fixes can be found here:
https://b.corp.google.com/issues/357913557#comment3

Fixes: 357913557
Test: Manually with splitting 2 running apps, and splitting and
      launching new one
Flag: EXEMPT bugfix
Change-Id: Ic8a5aeecbddef4ce0ea7a648a166c5430ceb3280

Former-commit-id: b9a8dd77348d2598d98f4f5b311962fbd7584983