[flexiglass] Changes SceneContainerViewModel.isVisible to snapshot state
Instead of a flow, it's now a snapshot state.

This CL includes changes needed in the view-binder to support that.

Bug: 354270224
Test: unit tests updated
Test: manually verified that scene container visibility is kept
up-to-date by unlocking (which makes it go away), pulling down the shade
(which makes it visible again), collapsing the shade (invisible), and
locking (visible).
Flag: com.android.systemui.scene_container

Change-Id: I6c9fd796940c2c4add98f7ad802805238b20143f