Add ManageWindowsViewContainer.
Adds an abstract view container that will be used for Manage Windows
menu on both Shell and Taskbar sides. This class will create the base
menu view and populate it with icons, given a list of TaskSnapshots
mapped to their respective RecentTaskInfo.

The child classes will be responsible for providing this map as well as
defining onClick behavior for the instance icons.

Bug: 336289597
Bug: 337903443
Test: Manual
Flag: com.android.window.flags.enable_desktop_windowing_multi_instance_features

Change-Id: I600cfdf3d20834f9917dee87f56159a4155d46e5

Former-commit-id: ea16e0e34a10ff355e7832c598fa295a0461877c