Handle communal transitions during other KTF transitions
This change handles the hub closing in the middle of another KTF
transition. In this case, we should cancel the existing transition and
transition immediately to the GLANCEABLE_HUB state, so that we can
properly transition away from it.

This causes KeyguardTransitionScenariosTest to fail, as that test uses a
FakeKeyguardTransitionRepository which behaves differently that a real
one - and doesn't properly support manual transitions. Decided to remove
some duplicate test cases there which are already covered by
CommunalSceneTransitionInteractorTest

Fixes: 360207879
Test: atest KeyguardTransitionScenariosTest
Test: atest CommunalSceneTransitionInteractorTest
Flag: com.android.systemui.communal_scene_ktf_refactor
Change-Id: Id2a00fa44dbf8508373da3ba276df6f591d41fd5