Relax delivery group policy constraints for prioritized bcasts.
When a receiver has been deferred for a prioritized broadcast,
it indicates that a decision has already been made to deliver
to other receivers in lower priority tranches. In that case,
upon the arrival of a new broadcast, the old broadcast for a
deferred receiver can be skipped and removed, provided that
the receiver is also an intended recipient of the new broadcast.

Bug: 344407893
Test: atest ./services/tests/mockingservicestests/src/com/android/server/am/BroadcastQueueTest.java
Test: atest ./services/tests/mockingservicestests/src/com/android/server/am/BroadcastQueueModernImplTest.java
Flag: NONE small fix
Change-Id: I386c823e58a7d586f5c8b6eace0ffd5f12ce16e8

Former-commit-id: 00d699f3a042010aa6777ca8015f8ef484d9d610