Move notifying userActivity to background thread in password textview
moving the the powermanager.userActivity call in password text view to
the background thread. this call signals to the system that the
useractivity has occured and in password text view this is called when
the text is appended/deleted/reset in the textview.

moving this call to onUserActivity listener impl in keyguardabskeyinputviewcontroller,
in which its called on background thread.

Test: checked perfetto binder transaction
Bug: 325203885
Flag: com.android.systemui.notify_password_text_view_user_activity_in_background

Change-Id: Id4533dca5452d03803a369a1026cae70e047c3c7

Former-commit-id: 33fd117c0ae0d8969fe0eb362bc9afe9dcc95bb7