Add a small delay for scheduling idle while sleeping
This is a follow up of [1]. Though that already covers most launch
duration, the twice switches of power mode may be unnecessary.
So add a small delay to check idle while device is sleeping and
there is a launching app, then if the device is unlocked before
the delay idle message reaches, the launch power mode can cover
entire launch operation.

[1]: I2383eaacdef9830a9b6fe638b2953f6a069f8d32

Bug: 332221147
Flag: EXEMPT bugfix
Test: Hot launch app from notification on lockscreen.
      Check trace StartModeLaunch ~ EndModeLaunch.
      The duration can cover most of "launching: $pkgName".
Change-Id: I723ce0da1434c6f69a77c10f19f7826b4f68dbde

Former-commit-id: 977f867e7454c54c4d9bbf13d1d851b78aba5397