Adding MSDL Feedback to the Pin bouncer.
MSDL haptic feedback is being added to interactions in the pin bouncer.
These include tapping the numpad keys, tapping and long-pressing the
delete key, and successful and failed authentication when the pin is
submitted using the return key. In addition, the tapping feedback is
added to the Emergency button.

Test: atest KeyguardAbsKeyInputViewControllerTest
Test: manual. Performing all interactions above and verifying that the
  correct haptics are delivered when the flag is on.
Flag: com.android.systemui.msdl_feedback
Bug: 361321945
Bug: 352600066
Change-Id: I481ae644070eb8da51530bdf7463d1c51fd8b7b5

Former-commit-id: 92ff273e9260cdcc542dfe5f94872fdfe6b13560