Refactor ensurePip call to be more explicit
The logic in moveToPinned (regarding requestStart and
ready) was half implicit based on controller state. This
makes it hard to understand when things will happen and
also makes it hard to build tracking logic.

This lifts the implicit logic (determining whether transition
is created and needs to requestStart) up to ensurePip where
the decision is actually made.

Bug: 325114242
Test: refactor, so existing tests
Flag: EXEMPT mechanical refactor
Change-Id: If406fe1590eacbc25199dc40cd460bcd7b17c813