Fix Chronometer countdown
Fixes: 359936006

The Chronometer countdown was counting down by 2 seconds rather
than by one second. This CL does two things: it fixes the
delay to work when counting down. It also rounds off the time
so that it doesn't count zero for two seconds. There is
special handling for counting down. The previous behavior started
with rounding down the count down. With the new rounding behavior,
count down has to do something different so that it achieves the same
behavior.

Flag: EXEMPT bugfix

Test: new test, manual testing
Change-Id: I611d21f1abb547d4ba5f2c3b269c1be1d480be0a

Former-commit-id: 7c6ba2e8ea73848250dc485462ebe4fcf08afccf