Ensure MaximizeMenu is released after close animation
Adds a local val to hold on to the MaximizeMenu ViewContainer reference
during the close animation, so that when the animation ends, the view
container can be released.
Prior to this change, the view container reference was set to null while
the animation ran, which meant that when it finished the attempt to call
release() silently failed because it was now null.

Bug: 362282037
Fix: 361843466
Flag: EXEMPT refactor
Test: open/close maximize menu multiple times, "dumpsys SurfaceFlinger"
doesn't show any Maximize Menu layer

Change-Id: I2947c58666bfb7e70e27f0df7c7247646ae70e4a

Former-commit-id: 9c209525c4e8a8d74d3cbbdf69c2e41700927111