Prevent dream starting on top of edit mode
Due to b/362830856, the dream may start over edit mode because we
briefly transition back to GLANCEABLE_HUB before transitioning away from
it again. This fix ignores that duplicate transition by only starting
the dream if we are on keyguard (eg not if the device is unlocked).

This is a short-term fix, the longer-term fix for b/362830856 is more
involved as it requires re-working the edit mode transition.

Fixes: 362719648
Test: atest CommunalDreamStartableTest
Flag: com.android.systemui.communal_hub
Change-Id: I1515bdc45ac3ae90434c6a7f05eb2bdfacab2f96

Former-commit-id: a2bb7b0fab63eab75af9ace929d58c4ade65ac94