Improve transitions when replacing an overlay
This CL improves transitions when replacing an overlay for elements that
are only in one of the overlays but also in the current scene. For these
elements, we now interpolate the state in the current scene with the
state in the overlay given the transition progress.

See b/353679003#comment35 for before/after videos.

Test: atest OverlayTest
Bug: 353679003
Flag: com.android.systemui.scene_container
Change-Id: Icebb02ddc60ae6a67c6cfd85d7d1bcb37613c507

Former-commit-id: 693b6465d1c7f38cc07e5060cc85092d174d0d69