Revert "Move showing keyguard after the UserSwitchObservers."
This reverts commit 0fbd151a3ad686de0afe1222b53c2b70971ac4a7 [formerly ad64f73046717374d4f9bc613307d01e93e19ebe].

Reason for revert: This CL caused b/360838273, we need revert it and show the keyguard at the beginning of the user switch as we were doing before. Otherwise, it can cause security issues such as the ones described in b/360838273#comment22.

This needs to be submitted together with the proper fix in `WindowManagerService.lockDeviceNow()`.

Bug: 331853529
Fixes: 360838273

Change-Id: I8f25255ac204160092fb88c9f099c697b222dc88

Former-commit-id: 8f012349dc728dda8a7d5ba9ae43fae0ecc7d322