[bc25] Make Lockscreen notifications content-key agnostic.
In the shade scenes implementation, we had to hide notifications on the
lockscreen when the notifications shade is open, in order to prevent a
double-composition of the same ElementKey in the same scene
(notifications in the shade and in the lockscreen, both of which were
rendered by the NotificationsShadeScene composable). With STL overlays,
this is no longer needed.

Bug: 356596436
Flag: com.android.systemui.scene_container
Flag: com.android.systemui.dual_shade
Test: Updated unit tests.
Test: Existing unit tests still pass.
Change-Id: I39bc95cd7fcda70b0e7ea36c59a714d4bf33fef0

Former-commit-id: 52917fc9d93b433e70ea121a7625531939e55b3c