Updated relevant aidl to pass current user when setting PhoneAcctSuggSrvc.
This CL updates the relevant Telecom Service aidl to ensure the current primary user can be passed when setting the PhoneAccountSuggestionService for Telecom CTS test purposes. In HSUM mode, this should be user 10 and in non HSUM mode this should be user 0. Without this change, PhoneAccountSuggestionServiceTest tests were failing in HSUM mode.

Test: atest PhoneAccountSuggestionServiceTest
Bug: 358587742
Flag: TEST_ONLY
Change-Id: I81431805da813c48ab2ef776e121a60b59abec63

Former-commit-id: 46461c142d32437c99776062682cf6c65fcaf9ab