[flexiglass] Extracts shadeMode from ShadeInteractor
Creates a smaller ShadeModeInteractor and extract the shadeMode and
isShadeLayoutWide to it then makes ShadeInteractor depend on it and
implement the extracted interface by delegation.

The CL also makes the two SceneResolver implementations that depended on
ShadeInteractor just for shadeMode use ShadeModeInteractor instead.

The second part is what breaks a runtime dependency cycle that was
making the CustomizationProvider content provider fail to be injected
with dependencies, leading to the attached bugs.

Fix: 361328659
Fix: 361382372
Test: unit tests extracted to new file/class
Test: manually verified that the lockscreen shortcuts and lockscreen
preview features in Wallpaper Picker work as intended
Test: manually verified that the exceptions in the AppComponentFactory
logcat tag are gone (were visible when starting sysui and when opening
wallpaper picker)
Flag: com.android.systemui.scene_container

Change-Id: Ic1218557867828ccacdb93384ef34d380acc3cb9

Former-commit-id: 79e3576e2d345f56637856d28f47053fcdf88f4b