Avoid starting window being stuck while launch transient launch.
While entering transient launch transition, the transient hide task and
it's descendant won't be add to sync group, so if an Activity request
finish itself during collect transient launch, the starting window can
be stuck to next activity because the close activity transition will be
collect within the same transition.
So while able to remove starting window, checking the sync state of the
activity instead of isCollecting.

Flag: EXEMPT bugfix
Bug: 365040473
Test: atest ActivityRecordTests
Test: simulate with sample app by following the description on issue.
Change-Id: I444802c434a3522a07fcb4d647a8a8008b441018

Former-commit-id: 4bcca6499e8261e2569bff739ba02ccabb62441e