Check device sleeping state for bouncer visibility
On reset when the device is going to sleep, the bouncer
may be asked to be visible for SIM unlocks. Do not attempt
to show the bouncer if the device is going to sleep. The dozing
signal can sometimes come too late.

Fixes: 351426938
Test: atest StatusBarKeyguardViewManagerTest
Test: manual - race condition on reboot, hit power button when on
SIM unlock
Flag: com.android.systemui.sim_pin_race_condition_on_restart

Change-Id: Id3ecac03bdaf31b8d54b63a44ef99a582a89a85b