[Improve logging] Add clear messaging for ignored logs
Improve logging by adding clear messaging when the
PresentationsStatsEventLogger would be ignoring logs.
Another option was to not log, and do an early return
in the case of when logging would be ignored. However,
I think that it may be helpful for debugging purpose
to still have logging to know if something useful is
being ignored.

Test: m
Flag: EXEMPT bugfix
Bug: 365801998
Change-Id: I377423d4dbd2ec50bae20a91436906009edbf191

Former-commit-id: d032a9c8d402a40facc32afc6e5567118258a1c8