Add buffer release callback in NDK
Introduce a new API, ASurfaceTransaction_setBufferWithRelease, which provides C/C++ code with the ability to register a callback function that is executed when a buffer is ready to be reused. This functionality mirrors the existing Java Transaction#setBuffer API, allowing for correct buffer management if the buffer is released when its overwritten in a transaction before the transaction is applied.

Flag: EXEMPT NDK
Test: atest ASurfaceControlTest
Bug:362513091

Change-Id: I49fd8d21adb34c193144035cfa2bee5bd6143dab

Former-commit-id: 3f6b1035c388e5db0232b9a4a4bf713f5df93058