Rename legacy cache keys
This CL renames legacy PropertyInvalidatedCache keys to match the new
naming convention, which is "cache_key.<module>.<name>", where the
module is one of "system_server", "telephony", or "bluetooth".  The
rename is entirely in the framework.  There is no backcompat issue.

Invalid cache keys are logged, unless the flag is true, in which case
a fatal exception is thrown.

Tested with a special image that removed the SELinux policies that
permitted the legacy property names to work.  Booted the image and
verified no selinux violations in logcat.  This means no caches were
using the legacy properties.

Test: booted the image and verified that no caches with non-standard
   names existed in the system.  Verification used the output of
   `dumpsys cacheinfo`.

Flag: EXEMPT refactor
Bug: 366552454
Change-Id: Icfc43980def04c8aefaa4097a38eabbc7ca3b054

Former-commit-id: ef881f533c9f0eff33728129f7c4d70abe56521a