Fix bubble notifications not being removed when permission revoked
When the notification permission is revoked, we don't get a ranking
update, instead the notifications get canceled with reason package
banned. The notification permission was introduced post-bubbles so
this is something that got missed I think, usually removals come
via the ranking updates (e.g. notif channel is stopped or bubble
option on that channel is removed would cause a ranking update).

Fix for this is to allow onEntryRemovals for package banned to remove
a bubble.

Flag: EXEMPT bug fix
Test: manual - add some bubbles, revoke the notification permission,
               observe that the bubbles are removed
Test: atest BubblesTest
Bug: 363118567
Change-Id: I51b3bb72429bfd6e04854c83c9f1c7d58d144bce

Former-commit-id: 1a399cda7570d2c1d0a5fc2ec2608cb9899dbfb6