Remove stateIn from buttons in shortcuts combined view model
This stateIn appears to be preventing the shortcuts from being updated
after the device is unlocked the first time. Adjusting the `started` to
`WhileSubscribed()` was adjusted to `Eagerly` to mitigate an issue with
wallet.

It is best to just remove this stateIn for now to retain functionality.
Further attempts to cut down on coroutine continuations can be made at
a later time.

Fixes: 366370507
Test: manual - reboot phone with shortcuts on LS, unlock and observe shortcuts show up
Flag: EXEMPT bugfix
Change-Id: I74e8e7751e938dd587a9ab435ebfb128723e0241