[TMP]: add dummy gradle-8.*-bin.zip
Google thought, it would be an awesome idea to add graddle bundled
as a zip into this repository. This is anyway an antipatter, but in
theis case the zip are bigger than 100mb, so github/gitlab forces
lfs, which is a very bad idea. The repo has been cleaned out
of those files and now to make a proper merge, we need to create a
dummy gradle-8.5-10-bin.zip, because in google sources, it is existent.

Signed-off-by: Julian Veit <claymore1298@gmail.com>